==> R/01_data_querying.Rmd <==
---
title: "01_data_querying"
author: "Markus Min"
date: "1/14/2022"
output: html_document
---

### Description
This Rmd file will go through querying PTAGIS and cleaning this data in order to prepare it for analysis.

# Determine datasets to query
To determine which datasets (natal origins) to include in our analysis, we are going to begin by examining all adult returns (detections in the adult fishway at BON) since 2005.

This first query will take the following attributes:

Advanced Reporting Home Page > Create Query Builder2 Report > Interrogation Summary

Once in Query Builder2, select the following filters:
- 1 Attributes: Standard + Release Site Basin, Release Site Subbasin, mark site basin, mark site subbasin, release site longitude, release site latitude, mark site, hatchery, stock name, mark capture method, release site code, mark site code, mark length, mark comment, mark conditional comment, mark data prjoect, mark data project name, mark session message
- 2 Metrics: Count, First Obs Date Max
- 16 First Obs Date/Time: 
  - From: Value 
  - Is: Greater than or equal to
    - Value: 6/1/2005 12:00:00 AM
- 18 Obs Site: BO1 - Bonneville Bradford Is. Ladder, BO2 - Bonneville Cascades Is. Ladder, BO3 - Bonneville WA Shore Ladder/AFF, BO4 - Bonneville WA Ladder Slots, B2A - BONNEVILLE ADULT WA SHORE, BWL - Bonneville WA Shore Ladder
  - These are all of the Bonneville adult fishways and therefore will allow you to identify all adult returns
- 25 Species: Steelhead

Run Report, and download as a CSV. Save this in PTAGIS_queries > intermediate_files > all_adult_returns.csv

## Initial query - v2
instead of an interrogation summary, let's run a tag detail query

Advanced Reporting Home Page > Create Query Builder2 Report > Tag Detail

- 1 Attributes: Standard + Hatchery, Mark Site Basin, Mark Site Subbasin, Release Site Basin, Release Site Subbasin, Stock, capture method
- 2 Metrics: Count, First Obs Date Max, Hatchery
- 16 Species: Steelhead
- 21 Mark Length: Less than or equal to 350 mm (to ensure you are only keeping individuals tagged as Juveniles)

These filters are too big for PTAGIS to handle. Split into multiple queries by release year
- even five years is too much; so try one year


## Inspect distribution of natal origins and run years
```{r load_libraries}
library(tidyverse)
library(here)
library(janitor)
library(lubridate)
```

```{r}
# Load table from Susannah on site info - investigate
relsites <- read.csv(file = here::here("covariate_data", "relsites.txt"), sep = "|", header = TRUE)

sort(unique(relsites$mrr_huc8name))
unique(relsites$mrr_huc6name)

subset(relsites, mrr_huc8name == "Lower Snake-Tucannon")

table(relsites$mrr_type)
```



```{r import_reformat_data}
adult_returns_complete <- clean_names(read.csv(here::here("PTAGIS_queries", "intermediate_files", "all_adult_returns.csv")))

# Remove all individuals where the mark length is greater than 350 mm, but make sure to keep NAs
subset(adult_returns_complete, mark_length_mm <= 350 | is.na(mark_length_mm)) -> adult_returns

# Also remove all individuals with comments that indicate they were captured as adults
table(adult_returns$mark_capture_method_name)
as.data.frame(table(subset(adult_returns, is.na(mark_length_mm))$mark_conditional_comment_name)) %>% 
  arrange(desc(Freq))

# Keywords to identify individuals tagged as adults: mature, adult, returning

# Look at mark_session_message_value
as.data.frame(table(subset(adult_returns, is.na(mark_length_mm))$mark_session_message_value)) %>% 
  arrange(desc(Freq))
# Here you might also be able to ID based on words like "adult" and "reproduction" in the project description, but these are less reliable

# Find all mark conditional comments that contain the words "mature", "adult", or "returning"
unique_comments <- unique(adult_returns$mark_conditional_comment_name)
adult_comments <- unique_comments[grep("mature|Mature|adult|Adult|returning|Returning", unique_comments)]
# Take out the word immature
adult_comments <- adult_comments[!(adult_comments %in% c("Immature", "Caudal Fin Damage, Immature"))]

# Remove all fish with comments indicating that they were tagged as adults
adult_returns_complete %>% 
  subset(., mark_length_mm <= 350 | is.na(mark_length_mm)) %>%  # keep those with NA mark lengths or mark lengths <= 350 mm
  subset(., !(mark_conditional_comment_name %in% adult_comments)) -> adult_returns # remove individuals known to have been tagged as adults


nrow(subset(adult_returns, rear_type_code == "H" & is.na(mark_length_mm)))
nrow(subset(adult_returns, rear_type_code == "H" & !(is.na(mark_length_mm))))
# Approx 50% of hatchery hatchery individuals are missing a mark length


# Keep only certain columns
adult_returns %>% 
  distinct(tag_code, .keep_all = TRUE) %>% 
  dplyr::rename(mark_site_code = mark_site_code_value, release_site_code = release_site_code_value,
                mark_comment = mark_comment_name) %>% 
  # dplyr::select(tag_code, first_time_value, rear_type_code, release_site_name, release_site_basin_name, release_site_subbasin_name,
  #               mark_site_basin_name, mark_site_subbasin_name, mark_site_name, hatchery_name, stock_name, mark_site_code, release_site_code,
  #               mark_comment, mark_length_mm) %>% 
  mutate(time = mdy_hms(first_time_value)) %>% 
  mutate(date = date(time)) %>% 
  dplyr::select(-c(first_time_value, time))-> adult_returns

# Sort into run years
run_year <- c("05/06", "06/07", "07/08", "08/09", "09/10", "10/11", "11/12", "12/13", "13/14", "14/15", "15/16", "16/17", "17/18", "18/19", "19/20", "20/21", "21/22")
run_year_start <- seq(ymd("2005-06-01"), ymd("2021-06-01"), by = "years")
run_year_end <- seq(ymd("2006-05-31"), ymd("2022-05-31"), by = "years")

run_year_df <- data.frame(run_year, run_year_start, run_year_end)

adult_returns %>% 
  mutate(dummy =TRUE) %>% 
  left_join(run_year_df %>% mutate(dummy=TRUE), by = "dummy") %>% 
  filter(run_year_start <= date, run_year_end >= date) %>% 
  select(-c(dummy, run_year_start, run_year_end)) -> adult_returns

# Export this file as metadata
write.csv(adult_returns, here::here("covariate_data", "tag_code_metadata.csv"), row.names = FALSE)
```

### Inspect distributions by run years and origins
Compare to Shelby's data - subset 05/06 to 14/15
```{r}
adult_returns %>% 
  subset(!(run_year %in% c("15/16", "16/17", "17/18", "18/19", "19/20", "20/21", "21/22"))) -> adult_returns_subset

# basin name is not specific enough
basin_table <- as.data.frame(table(adult_returns_subset$release_site_basin_name))
subbasin_table <- as.data.frame(table(adult_returns_subset$release_site_subbasin_name))
subbasin_table %>% 
  dplyr::rename(subbasin = Var1, count = Freq) %>% 
  arrange(desc(count)) -> subbasin_table

basin_table %>% 
  dplyr::rename(subbasin = Var1, count = Freq) %>% 
  arrange(desc(count)) -> basin_table

subset(adult_returns, release_site_subbasin_name == "Lower Snake-Tucannon")

table(subset(adult_returns, release_site_subbasin_name == "Lower Snake-Tucannon")$release_site_name)

subbasin_table
basin_table

# Get name of all unique release sites
release_sites <- unique(adult_returns$release_site_name)
# Get names of all unique mark sites
mark_sites <- unique(adult_returns$mark_site_name)
```

### Hood River
Shelby: 1,828 hatchery, 278 wild
Me: 1,818 hatchery, 276 wild
```{r}
# Hood River
subset(adult_returns_subset, release_site_subbasin_name == "Middle Columbia-Hood")

# This includes a lot of fish from the Wind River, which is not part of the Hood River system

# look at a table of release site names
as.data.frame(table(subset(adult_returns_subset, release_site_subbasin_name == "Middle Columbia-Hood")$release_site_name))
# compare to mark sites
as.data.frame(table(subset(adult_returns_subset, mark_site_subbasin_name == "Middle Columbia-Hood")$mark_site_name))

# what are the fish that were released between Bonneville and John Day?
subset(adult_returns_subset, release_site_name == "COLR4 - Columbia River - Bonneville Dam to John Day Dam (km 234-347)")
# Lots were marked at the John Day dam, but unclear where they are from


### Inspect release sites

# Have to add Parkdale Hatchery, which is on the middle fork of the Hood River
hood_sites <- c(release_sites[grep("Hood", release_sites)], "PARK - Parkdale Hatchery")
# Tons of these fish came from the Oak Springs Hatchery on the Deschutes

hood_fish <-  subset(adult_returns_subset, release_site_name %in% hood_sites)
table(hood_fish$rear_type_code)
# 1,818 hatchery and 276 wild. Very close

```

### Fifteenmile Creek
Shelby: 352
Me: 344
- pretty close!
```{r}
fifteenmile_sites <- release_sites[grep("Fifteen", release_sites)]

fifteenmile_fish <-  subset(adult_returns_subset, release_site_name %in% fifteenmile_sites)
table(fifteenmile_fish$run_year)
```

### Deschutes River
Shelby: 807 wild
Me: 804 wild
- pretty close, maybe some of these are iteroparous individuals that go into multiple run years
```{r}
deschutes_sites <- release_sites[grep("Deschutes", release_sites)]

deschutes_fish <-  subset(adult_returns_subset, release_site_name %in% deschutes_sites)
# 776 wild

# What about mark site basin
deschutes_fish <-  subset(adult_returns_subset, mark_site_basin_name == "Deschutes")
# Tons of hatchery fish that weren't released in the Deschutes

# What if we take only those wild fish
deschutes_fish <-  subset(adult_returns_subset, mark_site_basin_name == "Deschutes" & rear_type_code == "W")
# 804 wild

# what about by release basin? Since Deschutes is a basin
deschutes_fish <-  subset(adult_returns_subset, release_site_basin_name == "Deschutes")
# 804 wild, 2 hatchery

# Keep only the wild
deschutes_fish <-  subset(adult_returns_subset, release_site_basin_name == "Deschutes" & rear_type_code == "W")
```

### John Day River
2119 wild fish vs 2121 fish from Shelby
I also had 2121 fish in my analysis - where did those fish go?
these 2 fish have an unknown release site, but were marked in the John Day. Manually added
```{r}
jdr_fish <- subset(adult_returns_subset, release_site_subbasin_name %in% c("Upper John Day", "North Fork John Day", "Middle Fork John Day", "Lower John Day"))

JDR_CTH_1 <- clean_names(read.csv(here::here("PTAGIS_queries", "complete_tag_histories", "john_day_river", "2022-01-14-john_day_river_CTH_2005_2015_1.csv")))
JDR_CTH_2 <- clean_names(read.csv(here::here("PTAGIS_queries", "complete_tag_histories", "john_day_river", "2022-01-14-john_day_river_CTH_2005_2015_2.csv")))

### Combine files, fix some column data types
JDR_CTH_1 %>% 
  bind_rows(., JDR_CTH_2) %>% 
  mutate(event_date_time_value = mdy_hms(event_date_time_value))-> JDR_CTH


setdiff(unique(JDR_CTH$tag_code), jdr_fish$tag_code)

subset(JDR_CTH, tag_code %in% c("3D9.1C2D814B91", "3D9.1C2D84C4F9"))

subset(adult_returns_subset, tag_code %in% c("3D9.1C2D814B91", "3D9.1C2D84C4F9"))

jdr_fish <- subset(adult_returns_subset, release_site_subbasin_name %in% c("Upper John Day", "North Fork John Day", "Middle Fork John Day", "Lower John Day") | tag_code %in% c("3D9.1C2D814B91", "3D9.1C2D84C4F9") )

# Release site name is "- Unknown"? What the heck? But we would find these if used mark site... but we also known that that's not super reliable.
```

### Umatilla River
Shelby: 496 hatchery, 470 wild
Me: 496 hatchery, 467 wild, 1 unknown
```{r}
# by release site name
umatilla_sites <- release_sites[grep("Umatilla", release_sites)]

umatilla_fish <-  subset(adult_returns_subset, release_site_name %in% umatilla_sites)

# by release site subbasin
umatilla_fish <- subset(adult_returns_subset, release_site_subbasin_name =="Umatilla")
table(umatilla_fish$rear_type_code)

subset(umatilla_fish, rear_type_code == "U")
# This unknown reartype code fish sure looks like a hatchery fish
```

### Walla Walla River
Shelby: 1,692 hatchery, 535 wild
Me: 1,685 hatchery, 531 wild
```{r}
# by release site name
wallawalla_sites <- release_sites[grep("Walla Walla", release_sites)]

wallawalla_fish <-  subset(adult_returns_subset, release_site_name %in% wallawalla_sites)
# Way short

# by release site subbasin
wallawalla_fish <- subset(adult_returns_subset, release_site_subbasin_name =="Walla Walla")
table(wallawalla_fish$release_site_name)
table(wallawalla_fish$rear_type_code)
subset(wallawalla_fish, rear_type_code == "H")
# 1.685 hatchery and 531 wild (pretty close!)

```


### Yakima River
Shelby: 302 wild
Me: 301 wild
```{r}
# by release site name
yakima_sites <- release_sites[grep("Yakima", release_sites)]

yakima_fish <-  subset(adult_returns_subset, release_site_name %in% yakima_sites)


# By release site basin
yakima_fish <- subset(adult_returns_subset, release_site_basin_name =="Yakima")
```


### Wenatchee River
Shelby: 4,147 hatchery, 310 wild
Me: 4,137 hatchery, 307 wild - pretty close
```{r}
# by release site name
wenatchee_sites <- release_sites[grep("Wenatchee", release_sites)]

wenatchee_fish <-  subset(adult_returns_subset, release_site_name %in% wenatchee_sites)


# By release site subbasin
wenatchee_fish <- subset(adult_returns_subset, release_site_subbasin_name == "Wenatchee")
table(wenatchee_fish$rear_type_code)

# Can we see the acclimation variable?
table(subset(wenatchee_fish, rear_type_code == "H")$release_site_name)
table(wenatchee_fish$release_site_name)
table(subset(wenatchee_fish, rear_type_code == "H")$stock_name)
table(subset(wenatchee_fish, rear_type_code == "H")$hatchery_name)

table(subset(wenatchee_fish, rear_type_code == "H")$mark_conditional_comment_name)
table(subset(wenatchee_fish, rear_type_code == "H")$mark_comment)

subset(wenatchee_fish, rear_type_code == "H" & stock_name == "")
```


### Entiat River
Shelby: 360 wild
Me: 357 wild
```{r}
# by release site name
entiat_sites <- release_sites[grep("Entiat", release_sites)]

# By release site subbasin - this is too broad
entiat_fish <- subset(adult_returns_subset, release_site_subbasin_name =="Upper Columbia-Entiat")
table(entiat_fish$rear_type_code)
# way too many fish, including some hatchery fish
# Most are hatchery releases at RRE and RIS, some wild releases at RRE and RIS as well

entiat_fish <-  subset(adult_returns_subset, release_site_name %in% entiat_sites)
# 357 wild fish - quite close
```


### Tucannon River
Shelby: 2,533 hatchery, 419 wild
Me: 2,525 hatchery, 415 wild
```{r}
# By release site subbasin
tucannon_fish <- subset(adult_returns_subset, release_site_subbasin_name == "Lower Snake-Tucannon")
table(tucannon_fish$rear_type_code)
# WAY too many fish - subbasin is not fine enough resolution
# Most are being released at Lower Granite

as.data.frame(table(tucannon_fish$release_site_name))
# I think we just want the Tucannon River itself

# by release site name
tucannon_sites <- release_sites[grep("Tucannon", release_sites)]

tucannon_fish <-  subset(adult_returns_subset, release_site_name %in% tucannon_sites)
table(tucannon_fish$rear_type_code)
# 2,525 hatchery, 415 wild. Very close!

table(subset(tucannon_fish, rear_type_code == "H")$hatchery_name)

```

### Clearwater River
Shelby: 3,175 hatchery, 1,163 wild
Me : 3175 hatchery, 1157 wild
```{r}
# by release site name
clearwater_sites <- release_sites[grep("Clearwater", release_sites)]

clearwater_fish <-  subset(adult_returns_subset, release_site_name %in% clearwater_sites)
table(clearwater_fish$rear_type_code)
# Way short on all fish

# By release site basin
clearwater_fish <- subset(adult_returns_subset, release_site_basin_name == "Clearwater")
table(clearwater_fish$rear_type_code)
# 3,175 hatchery fish, 1,157 wild
```

### Grande Ronde River
Shelby: 4,239 hatchery, 549 wild
Me: 4,236 hatchery, 545 wild
```{r}
# by release site name
granderonde_sites <- release_sites[grep("Grande Ronde", release_sites)]

granderonde_fish <-  subset(adult_returns_subset, release_site_name %in% granderonde_sites)
table(granderonde_fish$rear_type_code)
table(granderonde_fish$release_site_subbasin_name)
# Way short on all fish

# By release site basin
granderonde_fish <- subset(adult_returns_subset, release_site_subbasin_name %in% c("Lower Grande Ronde", "Upper Grande Ronde", "Wallowa"))
table(granderonde_fish$rear_type_code)
# 4,236 hatchery and 545 wild
```

### Salmon River
Shelby: 7,661 hatchery, 831 wild
Me: 7,655 hatchery, 827 wild
```{r}
# by release site name
salmon_sites <- release_sites[grep("Salmon", release_sites)]

salmon_fish <-  subset(adult_returns_subset, release_site_name %in% salmon_sites)
table(salmon_fish$rear_type_code)
# Way short on all fish

# By release site basin
salmon_fish <- subset(adult_returns_subset, release_site_basin_name == "Salmon")
table(salmon_fish$rear_type_code)
# 7655 hatchery, 827 wild, 8 unknown

# look into unknown origin fish
subset(salmon_fish, rear_type_code == "U")
# Nothing in here to indicate whether they were hatchery or wild
```

### Imnaha River
Shelby: 2,613 hatchery, 925 wild
Me: 2,610 hatchery, 921 wild
```{r}
# by release site name
imnaha_sites <- release_sites[grep("Imnaha", release_sites)]

imnaha_fish <-  subset(adult_returns_subset, release_site_name %in% imnaha_sites)
table(imnaha_fish$rear_type_code)
# 116 hatchery (way short), 918 wild (close)

# By release site subbasin
imnaha_fish <- subset(adult_returns_subset, release_site_subbasin_name == "Imnaha")
table(imnaha_fish$rear_type_code)
# 2610 hatchery, 921 wild
```




### Other sites
```{r}
# compare full dataset to Shelby's
as.data.frame(table(adult_returns$release_site_subbasin_name)) # full dataset
as.data.frame(table(adult_returns_subset$release_site_subbasin_name)) %>% 
  arrange(desc(Freq))

# New candidates for inclusion: Klickitat, Lochsa, Methow, Okanogan, Trout (or is this part of another river system?)

# Klickitat
klickitat_fish <- subset(adult_returns, release_site_subbasin_name == "Klickitat")
table(klickitat_fish$rear_type_code) 
table(subset(klickitat_fish, rear_type_code == "H")$run_year)
table(klickitat_fish$run_year)

# Lochsa
lochsa_fish <- subset(adult_returns, release_site_subbasin_name == "Lochsa")
table(lochsa_fish$rear_type_code) 
table(lochsa_fish$run_year)

# Methow
methow_fish <- subset(adult_returns, release_site_subbasin_name == "Methow")
table(methow_fish$rear_type_code) 
table(methow_fish$release_site_name)
table(methow_fish$run_year)
table(subset(methow_fish, rear_type_code == "W")$run_year)

# Okanogan
okanogan_fish <- subset(adult_returns, release_site_subbasin_name == "Okanogan")
table(okanogan_fish$rear_type_code) 
table(okanogan_fish$run_year)

# Trout
trout_fish <- subset(adult_returns, release_site_subbasin_name == "Trout")
table(trout_fish$rear_type_code) 
table(trout_fish$run_year)

deschutes_fish

# Upper columbia, Priest rapids
UCPA_fish <- subset(adult_returns, release_site_subbasin_name == "Upper Columbia-Priest Rapids")
as.data.frame(table(UCPA_fish$release_site_name))
# Ringold hatchery, right on the mainstem. Not a tributary

# Upper Salmon and Little salmon are part of the Salmon

# Middle Columbia - Hood
MCH_fish <- subset(adult_returns, release_site_subbasin_name == "Middle Columbia-Hood")
as.data.frame(table(MCH_fish$release_site_name))
# Mostly Hood, also Fifteenmile (already covered). But Wind River and Trout Creek (trib to Wind River) might be a pop?

wind_fish <- subset(adult_returns, release_site_name %in% c("WIND2R - Wind River, Washington", "TROUTC - Trout Creek (trib. to Wind River, Wash.)"))
table(wind_fish$rear_type_code) # 1070 wild
table(wind_fish$run_year)
table(wind_fish$release_site_name)

# Hells Canyon
HC_fish <- subset(adult_returns, release_site_subbasin_name == "Hells Canyon")
table(HC_fish$rear_type_code) 
table(HC_fish$run_year)
as.data.frame(table(HC_fish$release_site_name))

# Lower Snake
LS_fish <- subset(adult_returns, release_site_subbasin_name == "Lower Snake")
table(LS_fish$rear_type_code) 
table(LS_fish$run_year)
as.data.frame(table(LS_fish$release_site_name))
# No, all in mainstem of Snake

# Lower Snake - Asotin
LS_fish <- subset(adult_returns, release_site_subbasin_name == "Lower Snake-Asotin")
table(LS_fish$rear_type_code) 
table(LS_fish$run_year)
as.data.frame(table(LS_fish$release_site_name))
as.data.frame(table(subset(LS_fish, rear_type_code == "H")$release_site_name))
as.data.frame(table(subset(LS_fish, rear_type_code == "W")$release_site_name))

table(subset(LS_fish, rear_type_code == "W")$run_year)

# Chief Joseph
CJ_fish <- subset(adult_returns, release_site_subbasin_name == "Chief Joseph")
table(CJ_fish$release_site_name)
# All from Well's Hatchery, mainstem Columbia. 

# Similkameen
SK_fish <- subset(adult_returns, release_site_subbasin_name == "Similkameen")
table(SK_fish$rear_type_code)  # Just hatchery
table(SK_fish$run_year) # Not enough run years (only 8) - ignore

# Middle Columbia-Lake Wallula
wallula_fish <- subset(adult_returns, release_site_subbasin_name == "Middle Columbia-Lake Wallula")
table(wallula_fish$run_year)
table(wallula_fish$rear_type_code)
table(wallula_fish$release_site_name)
# Nope, mostly released in the mainstem
```
Here we need to check these counts against when the arrays were active: http://cbr.washington.edu/dart/pit_basin_years_all

*Methow River* - LMR (Lower Methot River) array came online March 2009 - so we can use 09/10 run year onwards. Wouldn't have been enough to fit Shelby's criteria, but we have plenty of fish now - over 2,000 hatchery, and around 300 wild fish.

*Klickitat River* - there are lots of hatchery fish, starting in 11/12. The first interrogation site came online in 2010, at White creek (tributary to Klickitat River). More promising though is the Lyle Falls Fishway (LFF) site, which came online in 2011. This is very close to the mouth, and might see all fish returning to the Klickitat. We would have over 2,000 hatchery fish starting in 12/13.

Lochsa River - part of the Clearwater. So ignore those since they're already part of another group.

*Okanogan River* - Over 1,000 hatchery fish. The site on the river closest to the mouth has only been active since 2013. There are some other interrogation sites in the basin that have been active since 2009. This one is borderline, might not have 8 years of reliable data, unless array was active by mid-2013. Array start date was 10/8/2013, so iffy

Trout River - part of Deschutes, so ignore.

*Wind River* - we do have over 1,000 wild fish here, from the Wind River and Trout Creek (trib to Wind River). There's an array at rkm 30 on the Wind River (active since 2012), and one on Trout Creek (active since 2007). Enough or no?

Hells Canyon - over 1000 hatchery fish, released at SNAKE4 (Snake River befre HCD). But I'm not seeing any interrogation sites there, so I guess we should ignore

*Lower Snake - Asotin* - Ignore the fish from SNKTRP, that's a dipper trap in the Snake. But there are about 540 wild fish from Asotin Creek, and there are interrogation sites available since 2009 (2011 is when the one at the mouth started). Would be enough fish.

Chief Joseph - ignore these, they're all released at Wells Hatchery, which is on the mainstem Columbia

# PTAGIS - query all data

### Get tag codes for all fish of interest

The tributary systems that we want to query are:

From Shelby:
1. Hood River
2. Fifteenmile Creek
3. Deschutes River
4. John Day River
5. Umatilla River
6. Walla Walla River
7. Yakima River
8. Wenatchee River
9. Entiat River
10. Tucannon River
11. Clearwater River
12. Grande Ronde River
13. Salmon River
14. Imnaha river

New tributaries:
15. Methow River
16. Klickitat River
17. Okanogan River
18. Wind River
19. Asotin Creek

#### Determine release site names for each tributary
```{r}
# Get name of all unique release sites
release_sites <- unique(adult_returns$release_site_name)

# Get release sites that correspond to each tributary

# Hood River
hood_river_sites <- c(release_sites[grep("Hood", release_sites)], "PARK - Parkdale Hatchery")

#  Fifteenmile Creek
fifteenmile_creek_sites <- release_sites[grep("Fifteen", release_sites)]

# Deschutes River
deschutes_river_sites <- unique(subset(adult_returns, release_site_basin_name == "Deschutes")$release_site_name)

# John Day River
john_day_river_sites <- unique(subset(adult_returns, release_site_subbasin_name %in% c("Upper John Day", "North Fork John Day", "Middle Fork John Day", "Lower John Day"))$release_site_name)

# Umatilla River
umatilla_river_sites <- unique(subset(adult_returns, release_site_subbasin_name == "Umatilla")$release_site_name)

# Walla Walla River
walla_walla_river_sites <- unique(subset(adult_returns, release_site_subbasin_name == "Walla Walla")$release_site_name)

# Yakima River
yakima_river_sites <- unique(subset(adult_returns, release_site_basin_name == "Yakima")$release_site_name)

# Wenatchee River
wenatchee_river_sites <- unique(subset(adult_returns, release_site_subbasin_name == "Wenatchee")$release_site_name)

# Entiat River
entiat_river_sites <- release_sites[grep("Entiat", release_sites)]

# Tucannon River
tucannon_river_sites <- release_sites[grep("Tucannon", release_sites)]

# Clearwater River
clearwater_river_sites <- unique(subset(adult_returns, release_site_basin_name == "Clearwater")$release_site_name)

# Grande Ronde River
grande_ronde_river_sites <- unique(subset(adult_returns, release_site_subbasin_name %in% c("Lower Grande Ronde", "Upper Grande Ronde", "Wallowa"))$release_site_name)

# Salmon River
salmon_river_sites <- unique(subset(adult_returns, release_site_basin_name == "Salmon")$release_site_name)

# Imnaha river
imnaha_river_sites <- unique(subset(adult_returns, release_site_subbasin_name == "Imnaha")$release_site_name)

# Methow River
methow_river_sites <- unique(subset(adult_returns, release_site_subbasin_name == "Methow")$release_site_name)

# Klickitat River
klickitat_river_sites <- unique(subset(adult_returns, release_site_subbasin_name == "Klickitat")$release_site_name)

# Okanogan River
okanogan_river_sites <- unique(subset(adult_returns, release_site_subbasin_name == "Okanogan")$release_site_name)

# Wind River
wind_river_sites <- c("WIND2R - Wind River, Washington", "TROUTC - Trout Creek (trib. to Wind River, Wash.)")

# Asotin Creek
asotin_creek_sites <- release_sites[grep("Asotin", release_sites)]
```

#### Export table of release sites + natal origin

```{r}
data.frame(natal_origin = c(
  rep("Hood_River", length(hood_river_sites)),
  rep("Fifteenmile_Creek", length(fifteenmile_creek_sites)),
  rep("Deschutes_River", length(deschutes_river_sites)),
  rep("John_Day_River", length(john_day_river_sites)),
  rep("Umatilla_River", length(umatilla_river_sites)),
  rep("Walla_Walla_River", length(walla_walla_river_sites)),
  rep("Yakima_River", length(yakima_river_sites)),
  rep("Wenatchee_River", length(wenatchee_river_sites)),
  rep("Entiat_River", length(entiat_river_sites)),
  rep("Tucannon_River", length(tucannon_river_sites)),
  rep("Clearwater_River", length(clearwater_river_sites)),
  rep("Grande_Ronde_River", length(grande_ronde_river_sites)),
  rep("Salmon_River", length(salmon_river_sites)),
  rep("Imnaha_River", length(imnaha_river_sites)),
  rep("Methow_River", length(methow_river_sites)),
  rep("Klickitat_River", length(klickitat_river_sites)),
  rep("Okanogan_River", length(okanogan_river_sites)),
  rep("Wind_River", length(wind_river_sites)),
  rep("Asotin_Creek", length(asotin_creek_sites))
),
release_site_name = c(hood_river_sites,
                 fifteenmile_creek_sites, 
                 deschutes_river_sites,
                 john_day_river_sites,
                 umatilla_river_sites,
                 walla_walla_river_sites,
                 yakima_river_sites,
                 wenatchee_river_sites,
                 entiat_river_sites,
                 tucannon_river_sites,
                 clearwater_river_sites,
                 grande_ronde_river_sites,
                 salmon_river_sites,
                 imnaha_river_sites,
                 methow_river_sites,
                 klickitat_river_sites,
                 okanogan_river_sites,
                 wind_river_sites,
                 asotin_creek_sites
                 )) -> origin_table

# Export the origin table
write.csv(origin_table, here::here("covariate_data", "natal_origin_table.csv"), row.names = FALSE)
```



#### Get tag codes for all fish in river systems
Here we will break these tag codes into files containing 5,000 unique tags. This will keep the complete tag history files under 100 MB so that they can be uploaded to GitHub, and should make PTAGIS smoother.
```{r}
adult_returns %>% 
  subset(., release_site_name %in% c(hood_river_sites, fifteenmile_creek_sites, deschutes_river_sites, john_day_river_sites, umatilla_river_sites, walla_walla_river_sites, yakima_river_sites, wenatchee_river_sites, entiat_river_sites, tucannon_river_sites, clearwater_river_sites, grande_ronde_river_sites, salmon_river_sites, imnaha_river_sites, methow_river_sites, klickitat_river_sites, okanogan_river_sites, wind_river_sites, asotin_creek_sites)) -> adult_returns_for_analysis

# Write tag codes to multiple text files
as.data.frame(unique(adult_returns_for_analysis$tag_code)) -> unique_tags
  
    ntags <- length(unique_tags[,1])
    nfiles <- ceiling(ntags/5000)
    
    # Make an empty list to store the different files
    file_list <- list()
    
    # Split the tags into multiple files
    for (i in 1:nfiles){
      file_list[[i]] <- unique_tags[(1 + 5000 * (i - 1)):(5000 * i),1][!(is.na(unique_tags[(1 + 5000 * (i - 1)):(5000 * i),1]))]
    }
    
    # export the files
    for (i in 1:nfiles){
      write.table(file_list[[i]], file = here::here("PTAGIS_queries", "intermediate_files", "tag_codes", paste0("tag_codes_", i, ".txt")), col.names = FALSE, row.names = FALSE, quote = FALSE) 
    }
```

For each of these, the reports "CTH_tag_codes_1/2/3/4/5/6/7" were run. Those reports have the following settings:
Query Builder2:
- 1 Attributes: Tag, Event Type, Event Site, Event Date, Event Release Site, Event Release Date, (these first six are all standard, the rest you have to add manually) Event Date Time, Event Site Basin, Event Site Info, Event Site Code, Event Site Subbasin, Event Site Latitude, Event Site Longitude, Event Site Info
- 2 Metrics: Standard
- 29 Tag Code: Import text file that you exported from R

Run 2: 7/25/22

For each of these, the reports "CTH_tag_codes_1/2/3/4/5/6/7" were run. Those reports have the following settings:
Query Builder2:
Report type: complete tag history
- 1 Attributes: Tag, Event Type, Event Site, Event Date, Event Release Site, Event Release Date, (these first six are all standard, the rest you have to add manually),
event date time antenna, antenna group, antenna group configuration
- 2 Metrics: Standard
- 29 Tag Code: Import text file that you exported from R



# OLD/DEFUNCT - PTAGIS workflow - single tributary
PTAGIS querying has been set up to work in two steps:


### Step 1 (in PTAGIS): Determine the tag codes of adult returns to Bonneville
On the PTAGIS website once you have logged in, navigate to Advanced Reporting Home Page > Create Query Builder2 Report > Interrogation Summary

Once in Query Builder2, select the following filters:
- 1 Attributes: Standard
- 2 Metrics: Count, First Obs Date Max
- 9 Mark Site Subbasin == Your river of interest
  - Example: If you were querying data for the John Day River, you could search for "John Day" and then select the subbasins. In this case, you would select Upper John Day, North Fork John Day, Middle Fork John Day, and Lower John Day.
- 16 First Obs Date/Time: 
  - From: Value 
  - Is: Between
  - Value: The first date in your query
  - And: The last date in your query
  - For example, if you wanted all fish from run years 05/06 to 14/15, using June 1 as the separating date, you would have:
    - Value: 6/1/2005 12:00:00 AM
    - And: 5/31/2015 11:59:59 PM
- 18 Obs Site: BO1 - Bonneville Bradford Is. Ladder, BO2 - Bonneville Cascades Is. Ladder, BO3 - Bonneville WA Shore Ladder/AFF, BO4 - Bonneville WA Ladder Slots, B2A - BONNEVILLE ADULT WA SHORE, BWL - Bonneville WA Shore Ladder
  - These are all of the Bonneville adult fishways and there::herefore will allow you to identify all adult returns
- 25 Species: Steelhead
- 29 Mark Length: Less than or equal to 350 (to ensure you are only keeping individuals tagged as Juveniles)

Run Report, and download as a CSV. Save this in PTAGIS_queries > intermediate_files

NOTE

### Step 2 (in R): Get the tag codes of the adult returns

We will write a function to extract all of the unique tags for adult returns and export this as a text file that you can then use to query PTAGIS to get complete tag histories.

Note: PTAGIS seems to struggle with queries of over 2,000 tag codes (even though it says it can take up to 15,000). This code breaks up your tag files into chunks of 2,000 in the case of large queries, so you will have to concatenate them later.

```{r identify_adult_return_tags}
# NOTE: DO NOT INCLUDE ".csv" and ".txt" in input_CSV or output_txt

ID_adult_returns <- function(input_CSV, output_txt, river_system){
  # Load the libraries
  library(tidyverse)
  library(here)
  library(janitor)
  
  BON_adult_returns <- clean_names(read.csv(here::here("PTAGIS_queries", "intermediate_files", river_system,  paste0(input_CSV, ".csv"))))
  
  as.data.frame(unique(BON_adult_returns$tag_code)) -> unique_tags
  
  if (length(unique_tags > 2000)){
    ntags <- length(unique_tags[,1])
    nfiles <- ceiling(ntags/2000)
    
    # Make an empty list to store the different files
    file_list <- list()
    
    # Split the tags into multiple files
    for (i in 1:nfiles){
      file_list[[i]] <- unique_tags[(1 + 2000 * (i - 1)):(2000 * i),1][!(is.na(unique_tags[(1 + 2000 * (i - 1)):(2000 * i),1]))]
    }
    
    # export the files
    for (i in 1:nfiles){
      write.table(file_list[[i]], file = here::here("PTAGIS_queries", "intermediate_files", river_system, paste0(output_txt, "_", i, ".txt")), col.names = FALSE, row.names = FALSE, quote = FALSE) 
    }
  }
  
  else {
   write.table(unique_tags, file = here::here("PTAGIS_queries", "intermediate_files", river_system, output_txt), col.names = FALSE, row.names = FALSE, quote = FALSE) 
  }
}
```



### Step 3 (in PTAGIS): Get complete tag histories for adult returns
- On the PTAGIS website once you have logged in, navigate to Advanced Reporting Home Page > Home > Create Query Builder2 Report > Complete Tag History

Query Builder2:
- 1 Attributes: Tag, Event Type, Event Site, Event Date, Event Release Site, Event Release Date, (these first six are all standard, the rest you have to add manually) Event Date Time, Event Site Basin, Event Site Info, Event Site Code, Event Site Subbasin, Event Site Latitude, Event Site Longitude
- 2 Metrics: Standard
- 29 Tag Code: Import text file(s) that you exported from R
  - NOTE: You may have to run multiple queries if you have more than 2,000 tag codes

# Individual Queries


## John Day River, Wild Steelhead


### Query 1 - adult returns to BON
Interrogation Summary

Query Builder2 filters:
- 1 Attributes: Standard
- 2 Metrics: Count, First Obs Date Max
- 9 Mark Site Subbasin == Upper John Day, North Fork John Day, Middle Fork John Day, and Lower John Day.
- 16 First Obs Date/Time: 
  - From: Value 
  - Is: Between
    - Value: 6/1/2005 12:00:00 AM
    - And: 5/31/2015 11:59:59 PM
- 18 Obs Site: BO1 - Bonneville Bradford Is. Ladder, BO2 - Bonneville Cascades Is. Ladder, BO3 - Bonneville WA Shore Ladder/AFF, BO4 - Bonneville WA Ladder Slots, B2A - BONNEVILLE ADULT WA SHORE, BWL - Bonneville WA Shore Ladder
- 25 Species: Steelhead
- 29 Mark Length: Less than or equal to 350

Export CSV file: 2022-01-14-john_day_river_bon_adult_returns_2005-2015.csv

#### Get adult tag IDs
```{r}
ID_adult_returns(input_CSV = "2022-01-14-john_day_river_bon_adult_returns_2005-2015", output_txt = "2022-01-14-john_day_river_bon_adult_returns_2005-2015_TAGS",
                 river_system = "john_day_river")
```

### Query 2 - complete tag histories
Complete Tag History

Query Builder2:
- 1 Attributes: Tag, Event Type, Event Site, Event Date, Event Release Site, Event Release Date, (these first six are all standard, the rest you have to add manually) Event Date Time, Event Site Basin, Event Site Info, Event Site Code, Event Site Subbasin
- 2 Metrics: Standard
- 29 Tag Code: Import text file that you exported from R
  - 2 files: 2022-01-14-john_day_river_bon_adult_returns_2005-2015_TAGS_1.txt AND 2022-01-14-john_day_river_bon_adult_returns_2005-2015_TAGS_1.txt
    - These are separate queries

Export CSV files, save in PTAGIS_queries -> complete_tag_histories-> john_day_river :
2022-01-14-john_day_river_CTH_2005_2015_1.csv
2022-01-14-john_day_river_CTH_2005_2015_2.csv



# Covariate queries
In addition to the PIT tag detection histories for fish, we also need to query the covariate data from CBR DART.

What Shelby did:

"Water temperature, flow, and spill data were downloaded on 5 January, 2017. Daily water temperatures in the forebay outflow of McNary, Ice Harbor, Lower Granite, Priest Rapids, Rocky Reach, and Wells dams from 2005 to 2015 were downloaded from Columbia River DART (www.cbr.washington.edu/dart). Records of spill and outflow during the months of January, February, and March were also downloaded from Columbia River DART for the same dams. Stream temperature data for each tributary were queried from the U.S. Geological Survey, U.S. Department of the Interior Bureau of Reclamation, and Washington Department of Ecology. Natal stream temperature data were obtained for the mainstem Umatilla, Walla Walla, and Tucannon rivers. Multiple years of stream temperature data from a mainstem site within 75 rkm of the river mouth were not available for the remaining four tributaries. Umatilla data were obtained from the U.S. Department of the Interior Bureau of Reclamation (https://www.usbr.gov). "

Regarding how to calculate mean temps:
"I calculated mean water temperature during the week each steelhead neared and potentially passed its home tributary. To do this, I first predicted when each steelhead approached home. PIT arrays were primarily located at dams on the Columbia and Snake Rivers, and not located in the immediate vicinity of tributary mouths. Therefore, for each tributary, I found the median number of days it took overshooting steelhead to move from the lower dam prior to their home tributary to the overshoot dam. Travel times were estimated between the lower dam and the overshoot dam because many steelhead appeared to overwinter in the mainstem river prior to entering and being detected in the natal stream. Median travel times to home stream detectors were much greater than to overshoot dams, and therefore did not reflect the time steelhead arrived near their tributary mouth. Travel times varied throughout the year, therefore median travel times were estimated on a month-by-month basis. Steelhead moved slower during the summer period of peak water temperatures, consistent with the findings of Keefer et al. (2009).

Next, I estimated mean water temperatures in the period steelhead likely neared their home tributary. By adding the monthly median travel time to the date each steelhead passed the dam prior to the home tributary. To account for variation in travel time, I created a week-long window, centered around the median monthly travel time, during which each steelhead was likely arrive near their home tributary. I found the average water temperature in the outflow of the upstream dam within the one-week window. For steelhead from the Umatilla, Walla Walla, and Tucannon rivers, I estimated the average water temperature in the home stream, and also found the average natal-mainstem water temperature difference by subtracting the mean temperature at the upstream dam from the mean temperature at the tributary monitoring station.


To navigate to the appropriate page on CBR DART:

Homepage > Basin conditions > Project all years

We then need to query each dam separately.

1. Select the dam
2. Select the data
  - Here, we would be interested in "Outflow KCFS, Daily Average", "Spill KCFS, Daily Average", "Temperature C, Daily Average", and maybe "Temperature C, Tailrace, Daily Average"
    - Need to clarify some of the differences

Dams: 
- BON-Bonneville
- DWR-Dworshak
- IHR-Ice Harbor
- JDA-John Day
- LGS-Little Goose
- LWG-Lower Granite
- LMN-Lower Monumental
- MCN-McNary
- PRD-Priest Rapids
- RIS-Rock Island
- RRH-Rocky Reach
- TDA-The Dalles
- WAN-Wanapum
- WEL-Wells
