==> R/20_covariate_data.Rmd <==
---
title: "Covariate Data and Movement Timing"
author: "Markus Min"
date: '2023-03-27'
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, message = FALSE, warning = FALSE)

# load libraries
library(here)
library(janitor)
library(tidyverse)
library(lubridate)
library(car)
library(ggthemes)

```

## Objectives

1. Assess concerns over data availability.
2. Investigate multicollinearity between our three continuous covariates: flow, spill, and temperature.
3. Look at passage timing, to determine when covariate data will be most useful.
4. Decide how covariate data will be matched to state occupancy (e.g., a window approach)


```{r functions}

# Function 1: reformat covariate data
cov_reformat <- function(input_data, variable_name){
  
  # need to deal with leap years
  leap_years <- seq(1960, 2024, by = 4)
  
  input_data %>% 
    pivot_longer(., cols = colnames(.)[2:ncol(.)]) %>% 
    dplyr::rename(year = name,
                  !!quo_name(variable_name) := value) %>% 
    mutate(year = gsub("X","",year)) %>% 
    mutate(year = as.numeric(year), day = as.numeric(day)) %>% 
    # drop day 366 in non-leap years
    dplyr::filter(!(day == 366 & !(year %in% leap_years))) %>%
    # Add a new column for date (combine year and day)
    mutate(date = format(as.Date(day, origin = paste0(year-1, "-12-31")))) %>% 
    arrange(year, day) -> output_data
  
  return(output_data)
}

# Function 2: combine covariate data into df with date, spill, flow, and temp columns
cov_combine <- function(flow_data, spill_data, temp_data){
  flow_data %>% 
  left_join(., dplyr::select(spill_data, spill, date), by = "date") %>% 
  left_join(., dplyr::select(temp_data, temp, date), by = "date") %>% 
  mutate(date = ymd(date)) %>% 
  # keep only 2004 onwards (years that have data for all cov)
  subset(date >= ymd("2004-01-01")) -> cont_cov
  
  return(cont_cov)
  
}


# Function 3: quantify data availability
cov_data_availability <- function(cont_cov){
  cont_cov %>% 
  mutate(data_available = ifelse(is.na(flow) & !(is.na(spill)) & !(is.na(temp)), "spill + temp",
                                 ifelse(is.na(flow) & is.na(spill) & !(is.na(temp)), "temp",
                                        ifelse(is.na(flow) & (is.na(spill)) & (is.na(temp)), "none",
                                               ifelse(!(is.na(flow)) & !(is.na(spill)) & !(is.na(temp)), "spill + temp + flow",
                                                      ifelse(!(is.na(flow)) & !(is.na(spill)) & (is.na(temp)), "spill + flow",
                                                             ifelse(!(is.na(flow)) & (is.na(spill)) & (is.na(temp)), "flow",
                                                                    ifelse((is.na(flow)) & !(is.na(spill)) & (is.na(temp)), "spill",
                                                                           ifelse(!(is.na(flow)) & (is.na(spill)) & !(is.na(temp)), "temp + flow",
                                                                                  "ERROR"))))))))) %>% 
  # make them factors for plotting
  mutate(data_available = fct_rev(factor(data_available, levels = c("spill + temp + flow", "spill + flow", 
                                                            "temp + flow", "spill + temp",
                                                            "temp", "flow", "spill", "none")))) -> cont_cov
  
  return(cont_cov)
}

# Function 3.5 - a different function to quantify just missing data
cov_data_availability_2 <- function(cont_cov) {
  cont_cov %>% 
    # Make a different column where it just shows missing data (this should be simpler) 
  mutate(temp_avail = ifelse(is.na(temp), NA, "temp")) %>% 
  mutate(flow_avail = ifelse(is.na(flow), NA, "flow")) %>% 
  mutate(spill_avail = ifelse(is.na(spill), NA, "spill")) %>% 
  dplyr::select(date, temp_avail, flow_avail, spill_avail) %>% 
  pivot_longer(cols = c(temp_avail, flow_avail, spill_avail))-> data_avail_df

  data_avail_plot <- ggplot(subset(data_avail_df, !(is.na(value))), aes(x = date, y = value)) +
    geom_point() +
    ylab("Data Availability")
  
  return(data_avail_plot)
}



# Function 4: Data availabilty visualization
cov_data_viz <- function(cont_cov) {
  
  
# okay, is there a pattern seasonally?
# data_avail_plot <- ggplot(cont_cov, aes(x = date, y = data_available)) +
#   geom_point() +
#   scale_x_date(date_breaks = "1 year", date_labels = "%Y") +
#   theme(panel.grid.minor = element_blank(),
#         axis.text = element_text(size = 7))
# 
# # v2 - missing data only
# data_avail_plot <- ggplot(cont_cov, aes(x = date, y = missing_data)) +
#   geom_point() +
#   scale_x_date(date_breaks = "1 year", date_labels = "%Y") +
#   theme(panel.grid.minor = element_blank(),
#         axis.text = element_text(size = 7))

temp_plot <- ggplot(cont_cov, aes(x = day, y = temp)) +
  geom_point() +
  scale_x_continuous(breaks = c(0,50,100,150,200,250,300,365))

flow_plot <- ggplot(cont_cov, aes(x = day, y = flow)) +
  geom_point() +
  scale_x_continuous(breaks = c(0,50,100,150,200,250,300,365))

spill_plot <- ggplot(cont_cov, aes(x = day, y = spill)) +
  geom_point() +
  scale_x_continuous(breaks = c(0,50,100,150,200,250,300,365))

# plot_list <- list(data_avail_plot, temp_plot, flow_plot, spill_plot)
plot_list <- list(temp_plot, flow_plot, spill_plot)

return(plot_list)

}

# Function 5: Plot regressions, return VIF and R^2
cov_regressions <- function(cont_cov){
  
# plot
flow_v_spill <- ggplot(cont_cov, aes(x = flow, y = spill)) +
  geom_point()

flow_v_temp <- ggplot(cont_cov, aes(x = flow, y = temp)) +
  geom_point()

temp_v_spill <- ggplot(cont_cov, aes(x = temp, y = spill)) +
  geom_point()

# reg_plots <- list(flow_v_spill, flow_v_temp, temp_v_spill)

# run regressions
spill_flow_lm <- lm(spill~flow, data = cont_cov)
spill_flow_lm_sum <- summary(spill_flow_lm)

temp_flow_lm <- lm(temp~flow, data = cont_cov)
temp_flow_lm_sum <- summary(temp_flow_lm)

spill_temp_lm <- lm(spill~temp, data = cont_cov)
spill_temp_lm_sum <-summary(spill_temp_lm)

return(list(flow_v_spill, spill_flow_lm_sum,
            flow_v_temp, temp_flow_lm_sum,
            temp_v_spill, spill_temp_lm_sum))
}

# Function 6: Variance inflation factors
cov_vif <- function(cont_cov){
  # run multiple linear regressions and calculate VIF
  spill_mlm <- lm(spill ~ flow + temp, data = cont_cov)
  spill_vif <- vif(spill_mlm)

  temp_mlm <- lm(temp ~ flow + spill, data = cont_cov)
  temp_vif <- vif(temp_mlm)

  flow_mlm <- lm(flow ~ spill + temp, data = cont_cov)
  flow_vif <- vif(flow_mlm)
  
  return(list(spill_vif, temp_vif, flow_vif))
} 

  

```



```{r load_data}
# BON
BON_flow <- read.csv(here::here("covariate_data", "BON", "basin_outflow_BON_allyears.csv"))[1:366,]
BON_flow_long <- cov_reformat(input_data = BON_flow, variable_name = "flow")
BON_spill <- read.csv(here::here("covariate_data", "BON", "basin_spill_BON_allyears.csv"))[1:366,]
BON_spill_long <- cov_reformat(input_data = BON_spill, variable_name = "spill")
BON_temp <- read.csv(here::here("covariate_data", "BON", "basin_tempc_tailrace_BON_allyears.csv"))[1:366,]
BON_temp_long <- cov_reformat(input_data = BON_temp, variable_name = "temp")
BON_temp_2 <- read.csv(here::here("covariate_data", "BON", "basin_tempc_BON_allyears.csv"))[1:366,]
BON_temp_2_long <- cov_reformat(input_data = BON_temp_2, variable_name = "temp")

# ICH
ICH_flow <- read.csv(here::here("covariate_data", "ICH", "basin_outflow_IHR_allyears.csv"))[1:366,]
ICH_flow_long <- cov_reformat(input_data = ICH_flow, variable_name = "flow")
ICH_spill <- read.csv(here::here("covariate_data", "ICH", "basin_spill_IHR_allyears.csv"))[1:366,]
ICH_spill_long <- cov_reformat(input_data = ICH_spill, variable_name = "spill")
ICH_temp <- read.csv(here::here("covariate_data", "ICH", "basin_tempc_tailrace_IHR_allyears.csv"))[1:366,]
ICH_temp_long <- cov_reformat(input_data = ICH_temp, variable_name = "temp")


# JDA
JDA_flow <- read.csv(here::here("covariate_data", "JDA", "basin_outflow_JDA_allyears.csv"))[1:366,]
JDA_flow_long <- cov_reformat(input_data = JDA_flow, variable_name = "flow")
JDA_spill <- read.csv(here::here("covariate_data", "JDA", "basin_spill_JDA_allyears.csv"))[1:366,]
JDA_spill_long <- cov_reformat(input_data = JDA_spill, variable_name = "spill")
JDA_temp <- read.csv(here::here("covariate_data", "JDA", "basin_tempc_tailrace_JDA_allyears.csv"))[1:366,]
JDA_temp_long <- cov_reformat(input_data = JDA_temp, variable_name = "temp")


# TDA
TDA_flow <- read.csv(here::here("covariate_data", "TDA", "basin_outflow_TDA_allyears.csv"))[1:366,]
TDA_flow_long <- cov_reformat(input_data = TDA_flow, variable_name = "flow")
TDA_spill <- read.csv(here::here("covariate_data", "TDA", "basin_spill_TDA_allyears.csv"))[1:366,]
TDA_spill_long <- cov_reformat(input_data = TDA_spill, variable_name = "spill")
TDA_temp <- read.csv(here::here("covariate_data", "TDA", "basin_tempc_tailrace_TDA_allyears.csv"))[1:366,]
TDA_temp_long <- cov_reformat(input_data = TDA_temp, variable_name = "temp")

# LGR
LGR_flow <- read.csv(here::here("covariate_data", "LGR", "basin_outflow_LWG_allyears.csv"))[1:366,]
LGR_flow_long <- cov_reformat(input_data = LGR_flow, variable_name = "flow")
LGR_spill <- read.csv(here::here("covariate_data", "LGR", "basin_spill_LWG_allyears.csv"))[1:366,]
LGR_spill_long <- cov_reformat(input_data = LGR_spill, variable_name = "spill")
LGR_temp <- read.csv(here::here("covariate_data", "LGR", "basin_tempc_tailrace_LWG_allyears.csv"))[1:366,]
LGR_temp_long <- cov_reformat(input_data = LGR_temp, variable_name = "temp")

# LMO
LMO_flow <- read.csv(here::here("covariate_data", "LMO", "basin_outflow_LMN_allyears.csv"))[1:366,]
LMO_flow_long <- cov_reformat(input_data = LMO_flow, variable_name = "flow")
LMO_spill <- read.csv(here::here("covariate_data", "LMO", "basin_spill_LMN_allyears.csv"))[1:366,]
LMO_spill_long <- cov_reformat(input_data = LMO_spill, variable_name = "spill")
LMO_temp <- read.csv(here::here("covariate_data", "LMO", "basin_tempc_tailrace_LMN_allyears.csv"))[1:366,]
LMO_temp_long <- cov_reformat(input_data = LMO_temp, variable_name = "temp")

# LGO
LGO_flow <- read.csv(here::here("covariate_data", "LGO", "basin_outflow_LGS_allyears.csv"))[1:366,]
LGO_flow_long <- cov_reformat(input_data = LGO_flow, variable_name = "flow")
LGO_spill <- read.csv(here::here("covariate_data", "LGO", "basin_spill_LGS_allyears.csv"))[1:366,]
LGO_spill_long <- cov_reformat(input_data = LGO_spill, variable_name = "spill")
LGO_temp <- read.csv(here::here("covariate_data", "LGO", "basin_tempc_tailrace_LGS_allyears.csv"))[1:366,]
LGO_temp_long <- cov_reformat(input_data = LGO_temp, variable_name = "temp")


# MCN
MCN_flow <- read.csv(here::here("covariate_data", "MCN", "basin_outflow_MCN_allyears.csv"))[1:366,]
MCN_flow_long <- cov_reformat(input_data = MCN_flow, variable_name = "flow")
MCN_spill <- read.csv(here::here("covariate_data", "MCN", "basin_spill_MCN_allyears.csv"))[1:366,]
MCN_spill_long <- cov_reformat(input_data = MCN_spill, variable_name = "spill")
MCN_temp <- read.csv(here::here("covariate_data", "MCN", "basin_tempc_tailrace_MCN_allyears.csv"))[1:366,]
MCN_temp_long <- cov_reformat(input_data = MCN_temp, variable_name = "temp")


# PRA
PRA_flow <- read.csv(here::here("covariate_data", "PRA", "basin_outflow_PRD_allyears.csv"))[1:366,]
PRA_flow_long <- cov_reformat(input_data = PRA_flow, variable_name = "flow")
PRA_spill <- read.csv(here::here("covariate_data", "PRA", "basin_spill_PRD_allyears.csv"))[1:366,]
PRA_spill_long <- cov_reformat(input_data = PRA_spill, variable_name = "spill")
PRA_temp <- read.csv(here::here("covariate_data", "PRA", "basin_tempc_tailrace_PRD_allyears.csv"))[1:366,]
PRA_temp_long <- cov_reformat(input_data = PRA_temp, variable_name = "temp")


# RIS
RIS_flow <- read.csv(here::here("covariate_data", "RIS", "basin_outflow_RIS_allyears.csv"))[1:366,]
RIS_flow_long <- cov_reformat(input_data = RIS_flow, variable_name = "flow")
RIS_spill <- read.csv(here::here("covariate_data", "RIS", "basin_spill_RIS_allyears.csv"))[1:366,]
RIS_spill_long <- cov_reformat(input_data = RIS_spill, variable_name = "spill")
RIS_temp <- read.csv(here::here("covariate_data", "RIS", "basin_tempc_tailrace_RIS_allyears.csv"))[1:366,]
RIS_temp_long <- cov_reformat(input_data = RIS_temp, variable_name = "temp")


# RRE
RRE_flow <- read.csv(here::here("covariate_data", "RRE", "basin_outflow_RRH_allyears.csv"))[1:366,]
RRE_flow_long <- cov_reformat(input_data = RRE_flow, variable_name = "flow")
RRE_spill <- read.csv(here::here("covariate_data", "RRE", "basin_spill_RRH_allyears.csv"))[1:366,]
RRE_spill_long <- cov_reformat(input_data = RRE_spill, variable_name = "spill")
RRE_temp <- read.csv(here::here("covariate_data", "RRE", "basin_tempc_tailrace_RRH_allyears.csv"))[1:366,]
RRE_temp_long <- cov_reformat(input_data = RRE_temp, variable_name = "temp")


# WEL
WEL_flow <- read.csv(here::here("covariate_data", "WEL", "basin_outflow_WEL_allyears.csv"))[1:366,]
WEL_flow_long <- cov_reformat(input_data = WEL_flow, variable_name = "flow")
WEL_spill <- read.csv(here::here("covariate_data", "WEL", "basin_spill_WEL_allyears.csv"))[1:366,]
WEL_spill_long <- cov_reformat(input_data = WEL_spill, variable_name = "spill")
WEL_temp <- read.csv(here::here("covariate_data", "WEL", "basin_tempc_tailrace_WEL_allyears.csv"))[1:366,]
WEL_temp_long <- cov_reformat(input_data = WEL_temp, variable_name = "temp")

# Some quick filtering to remove obviously incorrect values

# A lot of issues with fall/winter temperature at RRE
RRE_temp_long %>% 
  # drop any rows where they're the same value as the next three or the previous three
  filter(!(temp == lag(temp) & temp == lag(temp, 2) & temp == lag(temp, 3) |
             temp == lead(temp) & temp == lead(temp, 2) & temp == lead(temp, 3))) -> RRE_temp_long




# Export this data for modeling
write.csv(BON_temp_long, here::here("covariate_data", "for_model", "BON_temp.csv"), row.names = FALSE)
write.csv(BON_temp_2_long, here::here("covariate_data", "for_model", "BON_forebay_temp.csv"), row.names = FALSE)
write.csv(JDA_temp_long, here::here("covariate_data", "for_model", "JDA_temp.csv"), row.names = FALSE)
write.csv(TDA_temp_long, here::here("covariate_data", "for_model", "TDA_temp.csv"), row.names = FALSE)
write.csv(MCN_temp_long, here::here("covariate_data", "for_model", "MCN_temp.csv"), row.names = FALSE)
write.csv(PRA_temp_long, here::here("covariate_data", "for_model", "PRA_temp.csv"), row.names = FALSE)
write.csv(RIS_temp_long, here::here("covariate_data", "for_model", "RIS_temp.csv"), row.names = FALSE)
write.csv(RRE_temp_long, here::here("covariate_data", "for_model", "RRE_temp.csv"), row.names = FALSE)
write.csv(WEL_temp_long, here::here("covariate_data", "for_model", "WEL_temp.csv"), row.names = FALSE)
write.csv(ICH_temp_long, here::here("covariate_data", "for_model", "ICH_temp.csv"), row.names = FALSE)
write.csv(LMO_temp_long, here::here("covariate_data", "for_model", "LMO_temp.csv"), row.names = FALSE)
write.csv(LGO_temp_long, here::here("covariate_data", "for_model", "LGO_temp.csv"), row.names = FALSE)
write.csv(LGR_temp_long, here::here("covariate_data", "for_model", "LGR_temp.csv"), row.names = FALSE)


# lets compare tailrace and non-tailrace temps at BON
dplyr::select(BON_temp_2_long, c(temp, date)) %>% 
  dplyr::rename(main_temp = temp) -> BON_temp_2_long_forjoin

BON_temp_long %>% 
  dplyr::rename(tailrace_temp = temp) %>% 
  dplyr::select(tailrace_temp, date) %>% 
  left_join(., BON_temp_2_long_forjoin, by = "date") -> BON_temp_comparison

ggplot(subset(BON_temp_comparison, !is.na(tailrace_temp) & !(is.na(main_temp))), aes(x = main_temp, y = tailrace_temp)) +
  geom_point()

# what's up with the weird observations?
BON_temp_comparison %>% 
  filter(abs(tailrace_temp - main_temp) > 3) -> BON_weird_temps
BON_weird_temps
# tailrace temps all look fishy to me here

# okay, this tailrace temp is definitely wrong
subset(BON_temp_comparison, date %in% c("2014-03-11", "2014-03-12", "2014-03-13", "2014-03-14", "2014-03-15"))

BON_temp_2_long %>% 
  group_by(year) %>% 
  summarise(annual_daily_average = mean(temp, na.rm = TRUE))
# CONFIRMED - this is forebay temperature
```

#### plot spill data availability
```{r}
library(viridis)
spill_by_year_plot <- function(data, dam_name){
  ggplot(subset(data, year >= 2005), aes(x = day, y = spill, color = year)) +
  # scale_color_tableau(palette = "Tableau 20") +
  scale_color_viridis() +
  geom_point() +
  ggtitle(dam_name)

}

spill_by_year_plot(data = BON_spill_long, dam_name = "Bonneville")
spill_by_year_plot(data = MCN_spill_long, dam_name = "McNary")
spill_by_year_plot(data = PRA_spill_long, dam_name = "Priest Rapids")
spill_by_year_plot(data = RIS_spill_long, dam_name = "Rock Island")
spill_by_year_plot(data = RRE_spill_long, dam_name = "Rocky Reach")
spill_by_year_plot(data = WEL_spill_long, dam_name = "Wells")
spill_by_year_plot(data = ICH_spill_long, dam_name = "Ice Harbor")
spill_by_year_plot(data = LGR_spill_long, dam_name = "Lower Granite")

```




#### calculate windows
```{r}
# so, I think that the most up to date files are in the same folders as the model runs
snake_adults_states_complete <- read.csv(here::here("stan_actual", "deteff_ESU_models", "500iter_3chain_runs", "snake", "snake_adults_states_complete.csv"), row.names = 1)
lowcol_adults_states_complete <- read.csv(here::here("stan_actual", "deteff_ESU_models", "lower_columbia", "lower_columbia_adults_states_complete.csv"), row.names = 1)
midcol_adults_states_complete <- read.csv(here::here("stan_actual", "deteff_ESU_models", "500iter_3chain_runs", "middle_columbia", "middle_columbia_adults_states_complete.csv"), row.names = 1)
uppcol_adults_states_complete <- read.csv(here::here("stan_actual", "deteff_ESU_models", "500iter_3chain_runs", "upper_columbia", "upper_columbia_adults_states_complete.csv"), row.names = 1)

# combine them all
snake_adults_states_complete %>% 
  bind_rows(., midcol_adults_states_complete) %>% 
  bind_rows(., uppcol_adults_states_complete) %>% 
  bind_rows(., lowcol_adults_states_complete) -> ASC

# now add tag code metadata, for natal origins
origin_numeric <- data.frame(natal_origin = c("Asotin_Creek", 
                                        "Clearwater_River",
                                        "Deschutes_River", 
                                        "Entiat_River", 
                                        "Fifteenmile_Creek", 
                                        "Grande_Ronde_River", 
                                        "Hood_River",
                                        "Imnaha_River",
                                        "John_Day_River", 
                                        "Methow_River", 
                                        "Okanogan_River", 
                                        "Salmon_River", 
                                        "Tucannon_River", 
                                        "Umatilla_River",
                                        "Walla_Walla_River",
                                        "Wenatchee_River", 
                                        "Yakima_River"),
                             natal_origin_numeric = seq(1,17,1))

origin_rear_actual <- read.csv(here::here("stan_actual", "origin_rear_actual.csv"), row.names = 1)
origin_rear_actual %>% 
  dplyr::rename(natal_origin_numeric = natal_origin) %>% 
  left_join(origin_numeric, by = "natal_origin_numeric") %>% 
  dplyr::select(tag_code_2, natal_origin) -> tag_code_origins

ASC %>% 
  left_join(., tag_code_origins, by = "tag_code_2") -> ASC

# also note which ESU they're from, for plotting
ESU_origins <- data.frame(natal_origin = unique(ASC$natal_origin),
                          ESU = c(rep("Snake", 6),
                                  rep("Middle Columbia", 6),
                                  rep("Upper Columbia", 4),
                                  "Lower Columbia"))

ASC %>% 
  left_join(., ESU_origins, by = "natal_origin") -> ASC

# a function that takes just one state, and finds the average amount of time to move out of that state

residence_time <- function(residence_state){
  
  # don't keep any with implicit time interpolated
  ASC %>% 
    filter(state == residence_state & tag_code_2 == lead(tag_code_2) & pathway != "implicit" & lead(pathway) != "implicit" |
             lag(state) == residence_state & tag_code_2 == lag(tag_code_2) & pathway != "implicit" & lag(pathway) != "implicit") %>% 
    mutate(date_time = ymd_hms(date_time)) -> one_state_df
  
  # make a data frame to record all of the transitions
  n_transitions <- nrow(one_state_df)/2
  passage_df <- data.frame(tag_code_2 = one_state_df$tag_code_2[seq(1,(nrow(one_state_df)-1),2)],
                           ESU = one_state_df$ESU[seq(1,(nrow(one_state_df)-1),2)],
                           natal_origin = one_state_df$natal_origin[seq(1,(nrow(one_state_df)-1),2)],
                           passage_time = NA)
  
  
  for (i in 1:nrow(passage_df)){
    passage_df$passage_time[i] <- one_state_df$date_time[(i*2)] - one_state_df$date_time[(i*2-1)]
    
  }
  
  return(passage_df)
  
}

main_mainstem_states <- c(
  "mainstem, BON to MCN",
  "mainstem, MCN to ICH or PRA",
  "mainstem, PRA to RIS",
  "mainstem, RIS to RRE",
  "mainstem, RRE to WEL",
  "mainstem, ICH to LGR")

list_quantiles <- list()

for (i in 1:length(main_mainstem_states)) {
  residence_df <- residence_time(residence_state = main_mainstem_states[i])
  
  # Look into windows - show: 1) median of the whole distribution, 
  # and what cutoff you'd need to capture 50%, 80%, and 95% of fish residence time
  quantiles <- quantile(residence_df$passage_time, probs = c(0.5, 0.8, 0.95))
  list_quantiles[[i]] <- quantiles

  
}

# make an actual table
quantiles_table <- data.frame(reach = main_mainstem_states,
                              t50 = c(list_quantiles[[1]][1],
                                       list_quantiles[[2]][1],
                                       list_quantiles[[3]][1],
                                       list_quantiles[[4]][1],
                                       list_quantiles[[5]][1],
                                       list_quantiles[[6]][1]),
                              t80 = c(list_quantiles[[1]][2],
                                       list_quantiles[[2]][2],
                                       list_quantiles[[3]][2],
                                       list_quantiles[[4]][2],
                                       list_quantiles[[5]][2],
                                       list_quantiles[[6]][2]),
                              t95 = c(list_quantiles[[1]][3],
                                       list_quantiles[[2]][3],
                                       list_quantiles[[3]][3],
                                       list_quantiles[[4]][3],
                                       list_quantiles[[5]][3],
                                       list_quantiles[[6]][3]))

# write.csv(quantiles_table, here::here("covariate_data", "reach_quantiles.csv"))
```


### export temperature data using window approach
```{r}
# Re-export, using reach-specific residence time windows
# quantiles <- read.csv(here::here("covariate_data", "reach_quantiles.csv"), row.names = 1)

# round the median to nearest whole day
quantiles_table %>% 
  mutate(median = round(t50)) -> quantiles_table

# match them using outflow

# use a function
window_temp <- function(temp_data, median_time){
  temp_data %>% 
  filter(!is.na(temp)) %>% 
  mutate(date = date(date)) %>% 
  mutate(window_temp = NA) %>% # let's start in 2005, since that's when our data starts
  subset(year >= 2005) -> temp_data

# loop to get all windows
for (i in 1:nrow(subset(temp_data, year <= 2022))){
  temp_data$window_temp[i] <- mean(subset(temp_data, date >= temp_data$date[i] & date <= temp_data$date[i] + days(x = median_time))$temp)
}
  
  # now, add back in NAs for all values that we don't have
  temp_data %>% 
    complete(date = seq(ymd("2005-01-01"), ymd("2022-12-31"), by = "days")) %>% 
    mutate(year = year(date), day = day(date)) -> temp_data
  
  return(temp_data)
}

# Bonneville Dam, for mainstem, BON to MCN reach
BON_window_temp <- window_temp(temp_data = BON_temp_2_long, median_time = subset(quantiles_table, reach == "mainstem, BON to MCN")$median)

# McNary Dam, for mainstem, MCN to ICH or PRA reach
MCN_window_temp <- window_temp(temp_data = MCN_temp_long, median_time = subset(quantiles_table, reach == "mainstem, MCN to ICH or PRA")$median)

# PRA Dam, for PRA to RIS reach
PRA_window_temp <- window_temp(temp_data = PRA_temp_long, median_time = subset(quantiles_table, reach == "mainstem, PRA to RIS")$median)

# RIS Dam, for RIS to RRE reach
RIS_window_temp <- window_temp(temp_data = RIS_temp_long, median_time = subset(quantiles_table, reach == "mainstem, RIS to RRE")$median)

# RRE Dam, for RRE to WEL reach
RRE_window_temp <- window_temp(temp_data = RRE_temp_long, median_time = subset(quantiles_table, reach == "mainstem, RRE to WEL")$median)

# Ice Harbor Dam, for ICH to LGR reach
ICH_window_temp <- window_temp(temp_data = ICH_temp_long, median_time = subset(quantiles_table, reach == "mainstem, ICH to LGR")$median)

# export all of these
write.csv(BON_window_temp, here::here("covariate_data", "for_model", "temp", "BON_window_temp.csv"), row.names = FALSE)
write.csv(MCN_window_temp, here::here("covariate_data", "for_model", "temp", "MCN_window_temp.csv"), row.names = FALSE)
write.csv(PRA_window_temp, here::here("covariate_data", "for_model", "temp", "PRA_window_temp.csv"), row.names = FALSE)
write.csv(RIS_window_temp, here::here("covariate_data", "for_model", "temp", "RIS_window_temp.csv"), row.names = FALSE)
write.csv(RRE_window_temp, here::here("covariate_data", "for_model", "temp", "RRE_window_temp.csv"), row.names = FALSE)
write.csv(ICH_window_temp, here::here("covariate_data", "for_model", "temp", "ICH_window_temp.csv"), row.names = FALSE)

```

### Export March days of spill as covariate
```{r}
march_spill_days <- function(spill_data){
  spill_data %>% 
    # keep only 2005 onwards
    subset(year >= 2005) %>% 
    mutate(date = ymd(date)) %>% 
    mutate(month = month(date)) %>% 
    mutate(any_spill = ifelse(spill > 0, "yes", "no")) %>% 
    group_by(year, month) %>% 
    count(any_spill == "yes") %>% 
    dplyr::rename(positive_spill = `any_spill == "yes"`) %>% 
    subset(positive_spill == TRUE & month == 3) %>% 
    dplyr::rename(march_spill_days = n) %>% 
    ungroup() %>% 
    complete(year = seq(2005,2022,1), fill = list(march_spill_days = 0)) %>% 
    dplyr::select(year, march_spill_days) -> march_spill_days
    # summarise(monthly_spill_days = count(any_spill == "yes"))
  
  return(march_spill_days)
}

BON_march_spill <- march_spill_days(spill_data = BON_spill_long)
MCN_march_spill <- march_spill_days(spill_data = MCN_spill_long)
PRA_march_spill <- march_spill_days(spill_data = PRA_spill_long)
RIS_march_spill <- march_spill_days(spill_data = RIS_spill_long)
RRE_march_spill <- march_spill_days(spill_data = RRE_spill_long)
WEL_march_spill <- march_spill_days(spill_data = WEL_spill_long)
ICH_march_spill <- march_spill_days(spill_data = ICH_spill_long)
LGR_march_spill <- march_spill_days(spill_data = LGR_spill_long)

write.csv(BON_march_spill, here::here("covariate_data", "for_model", "spill", "BON_march_spill.csv"), row.names = FALSE)
write.csv(MCN_march_spill, here::here("covariate_data", "for_model", "spill", "MCN_march_spill.csv"), row.names = FALSE)
write.csv(PRA_march_spill, here::here("covariate_data", "for_model", "spill", "PRA_march_spill.csv"), row.names = FALSE)
write.csv(RIS_march_spill, here::here("covariate_data", "for_model", "spill", "RIS_march_spill.csv"), row.names = FALSE)
write.csv(RRE_march_spill, here::here("covariate_data", "for_model", "spill", "RRE_march_spill.csv"), row.names = FALSE)
write.csv(WEL_march_spill, here::here("covariate_data", "for_model", "spill", "WEL_march_spill.csv"), row.names = FALSE)
write.csv(ICH_march_spill, here::here("covariate_data", "for_model", "spill", "ICH_march_spill.csv"), row.names = FALSE)
write.csv(LGR_march_spill, here::here("covariate_data", "for_model", "spill", "LGR_march_spill.csv"), row.names = FALSE)

# make a quick plot for these
ggplot(MCN_march_spill, aes(x = year, y = march_spill_days)) +
  geom_point() +
  ggtitle("McNary Dam")

ggplot(PRA_march_spill, aes(x = year, y = march_spill_days)) +
  geom_point() +
  ggtitle("Priest Rapids Dam")

ggplot(RIS_march_spill, aes(x = year, y = march_spill_days)) +
  geom_point() +
  ggtitle("Rock Island Dam")

ggplot(RRE_march_spill, aes(x = year, y = march_spill_days)) +
  geom_point() +
  ggtitle("Rocky Reach Dam")

ggplot(WEL_march_spill, aes(x = year, y = march_spill_days)) +
  geom_point() +
  ggtitle("Wells Dam")

ggplot(ICH_march_spill, aes(x = year, y = march_spill_days)) +
  geom_point() +
  ggtitle("Ice Harbor Dam")

ggplot(LGR_march_spill, aes(x = year, y = march_spill_days)) +
  geom_point() +
  ggtitle("Lower Granite Dam")



```




```{r load_states_data}
# so, I think that the most up to date files are in the same folders as the model runs
snake_adults_states_complete <- read.csv(here::here("stan_actual", "deteff_ESU_models", "500iter_3chain_runs", "snake", "snake_adults_states_complete.csv"), row.names = 1)
lowcol_adults_states_complete <- read.csv(here::here("stan_actual", "deteff_ESU_models", "lower_columbia", "lower_columbia_adults_states_complete.csv"), row.names = 1)
midcol_adults_states_complete <- read.csv(here::here("stan_actual", "deteff_ESU_models", "500iter_3chain_runs", "middle_columbia", "middle_columbia_adults_states_complete.csv"), row.names = 1)
uppcol_adults_states_complete <- read.csv(here::here("stan_actual", "deteff_ESU_models", "500iter_3chain_runs", "upper_columbia", "upper_columbia_adults_states_complete.csv"), row.names = 1)

# combine them all
snake_adults_states_complete %>% 
  bind_rows(., midcol_adults_states_complete) %>% 
  bind_rows(., uppcol_adults_states_complete) %>% 
  bind_rows(., lowcol_adults_states_complete) -> ASC

# now add tag code metadata, for natal origins
origin_numeric <- data.frame(natal_origin = c("Asotin_Creek", 
                                        "Clearwater_River",
                                        "Deschutes_River", 
                                        "Entiat_River", 
                                        "Fifteenmile_Creek", 
                                        "Grande_Ronde_River", 
                                        "Hood_River",
                                        "Imnaha_River",
                                        "John_Day_River", 
                                        "Methow_River", 
                                        "Okanogan_River", 
                                        "Salmon_River", 
                                        "Tucannon_River", 
                                        "Umatilla_River",
                                        "Walla_Walla_River",
                                        "Wenatchee_River", 
                                        "Yakima_River"),
                             natal_origin_numeric = seq(1,17,1))

origin_rear_actual <- read.csv(here::here("stan_actual", "origin_rear_actual.csv"), row.names = 1)
origin_rear_actual %>% 
  dplyr::rename(natal_origin_numeric = natal_origin) %>% 
  left_join(origin_numeric, by = "natal_origin_numeric") %>% 
  dplyr::select(tag_code_2, natal_origin) -> tag_code_origins

ASC %>% 
  left_join(., tag_code_origins, by = "tag_code_2") -> ASC

```

## Matching covariate data to state occupancy

Example of movements for one fish:
```{r}
# subset(ASC, tag_code %in% c("384.3B2395F9F6", "3DD.0077D43358", "3DD.0077D3F3CB", "3DD.0077996B66", "3DD.00777260CF", "3DD.00776DF38F", "3DD.0077645B38", "3DD.0077437BC1", "3DD.00774166EC", "3D9.1C2DF4F5E2", "3D9.1C2DE0F9A7"))
## Matching covariate data to state occupancy
dplyr::select(subset(ASC, tag_code %in% c("384.3B2395F9F6")), state, date_time, pathway)
```



## Covariate Data

### Summary
- The following dams lack winter (November - Feburary) temperature data: Bonneville, Rock Island (until 2012), Rocky Reach (until 2012), Wells Dam (until 2013) 
- Spill data can be a bit patchy - are missing spill values zeros?
- Spill and flow are strongly correlated; VIF vary, but are between 2 and 8. R^2 values range 0.55 - 0.9
- In some years, spill appears to follow the relationship *spill = flow - X*, but the value of *X* changes from year to year. This appears to be especially true for the four Snake River Dams





### Bonneville Dam

#### Data availability
Availability of spill/temperature/flow data, by number of days (2004-present):
```{r BON}
BON_cont_cov <- cov_combine(flow_data = BON_flow_long,
                            spill_data = BON_spill_long,
                            temp_data = BON_temp_long)

BON_cont_cov <- cov_data_availability(BON_cont_cov)

# data availability as a table
table(BON_cont_cov$data_available)
```

#### Data by day of year
```{r}
cov_data_availability_2(cont_cov = BON_cont_cov)
# visualize data as plots
cov_data_viz(cont_cov = BON_cont_cov)
```


#### Regressions and VIF
``` {r}
# regressions and vif
cov_regressions(cont_cov = BON_cont_cov)
```
Variance inflation factors (Spill, Temperature, and Flow):

```{r BON_vif}
cov_vif(cont_cov = BON_cont_cov)
```



### John Day Dam

#### Data availability
Availability of spill/temperature/flow data, by number of days (2004-present):
```{r JDA}
JDA_cont_cov <- cov_combine(flow_data = JDA_flow_long,
                            spill_data = JDA_spill_long,
                            temp_data = JDA_temp_long)

JDA_cont_cov <- cov_data_availability(JDA_cont_cov)

# data availability as a table
table(JDA_cont_cov$data_available)
```

#### Data by day of year
```{r}
cov_data_availability_2(cont_cov = JDA_cont_cov)
# visualize data as plots
cov_data_viz(cont_cov = JDA_cont_cov)
```


#### Regressions and VIF
``` {r}
# regressions and vif
cov_regressions(cont_cov = JDA_cont_cov)
```
Variance inflation factors (Spill, Temperature, and Flow):
  
```{r JDA_vif}
cov_vif(cont_cov = JDA_cont_cov)
```



### The Dalles Dam

#### Data availability
Availability of spill/temperature/flow data, by number of days (2004-present):
```{r TDA}
TDA_cont_cov <- cov_combine(flow_data = TDA_flow_long,
                            spill_data = TDA_spill_long,
                            temp_data = TDA_temp_long)

TDA_cont_cov <- cov_data_availability(TDA_cont_cov)

# data availability as a table
table(TDA_cont_cov$data_available)
```

#### Data by day of year
```{r}
cov_data_availability_2(cont_cov = TDA_cont_cov)
# visualize data as plots
cov_data_viz(cont_cov = TDA_cont_cov)
```


#### Regressions and VIF
``` {r}
# regressions and vif
cov_regressions(cont_cov = TDA_cont_cov)
```
Variance inflation factors (Spill, Temperature, and Flow):
  
```{r TDA_vif}
cov_vif(cont_cov = TDA_cont_cov)
```


### McNary Dam

#### Data availability
Availability of spill/temperature/flow data, by number of days (2004-present):
```{r MCNcont_cov <- cov_combine(flow_data = MCN_flow_long,
                            spill_data = MCN_spill_long,
                            temp_data = MCN_temp_long)

MCN_cont_cov <- cov_data_availability(MCN_cont_cov)

# data availability as a table
table(MCN_cont_cov$data_available)
```

#### Data by day of year
```{r}
cov_data_availability_2(cont_cov = MCN_cont_cov)
# visualize data as plots
cov_data_viz(cont_cov = MCN_cont_cov)
```


#### Regressions and VIF
``` {r}
# regressions and vif
cov_regressions(cont_cov = MCN_cont_cov)
```
Variance inflation factors (Spill, Temperature, and Flow):
  
```{r MCN_vif}
cov_vif(cont_cov = MCN_cont_cov)
```



### Priest Rapids Dam

#### Data availability
Availability of spill/temperature/flow data, by number of days (2004-present):
```{r PRA}
PRA_cont_cov <- cov_combine(flow_data = PRA_flow_long,
                            spill_data = PRA_spill_long,
                            temp_data = PRA_temp_long)

PRA_cont_cov <- cov_data_availability(PRA_cont_cov)

# data availability as a table
table(PRA_cont_cov$data_available)
```

#### Data by day of year
```{r}
cov_data_availability_2(cont_cov = PRA_cont_cov)
# visualize data as plots
cov_data_viz(cont_cov = PRA_cont_cov)
```


#### Regressions and VIF
``` {r}
# regressions and vif
cov_regressions(cont_cov = PRA_cont_cov)
```
Variance inflation factors (Spill, Temperature, and Flow):
  
```{r PRA_vif}
cov_vif(cont_cov = PRA_cont_cov)
```



### Rock Island Dam

#### Data availability
Availability of spill/temperature/flow data, by number of days (2004-present):
```{r RIS}
RIS_cont_cov <- cov_combine(flow_data = RIS_flow_long,
                            spill_data = RIS_spill_long,
                            temp_data = RIS_temp_long)

RIS_cont_cov <- cov_data_availability(RIS_cont_cov)

# data availability as a table
table(RIS_cont_cov$data_available)
```

#### Data by day of year
```{r}
cov_data_availability_2(cont_cov = RIS_cont_cov)
# visualize data as plots
cov_data_viz(cont_cov = RIS_cont_cov)
```


#### Regressions and VIF
``` {r}
# regressions and vif
cov_regressions(cont_cov = RIS_cont_cov)
```
Variance inflation factors (Spill, Temperature, and Flow):
  
```{r RIS_vif}
cov_vif(cont_cov = RIS_cont_cov)
```


### Rocky Reach Dam

#### Data availability
Availability of spill/temperature/flow data, by number of days (2004-present):
```{r RRE}
RRE_cont_cov <- cov_combine(flow_data = RRE_flow_long,
                            spill_data = RRE_spill_long,
                            temp_data = RRE_temp_long)

RRE_cont_cov <- cov_data_availability(RRE_cont_cov)

# data availability as a table
table(RRE_cont_cov$data_available)
```

#### Data by day of year
```{r}
cov_data_availability_2(cont_cov = RRE_cont_cov)
# visualize data as plots
cov_data_viz(cont_cov = RRE_cont_cov)
```


#### Regressions and VIF
``` {r}
# regressions and vif
cov_regressions(cont_cov = RRE_cont_cov)
```
Variance inflation factors (Spill, Temperature, and Flow):
  
```{r RRE_vif}
cov_vif(cont_cov = RRE_cont_cov)
```


### Wells Dam

#### Data availability
Availability of spill/temperature/flow data, by number of days (2004-present):
```{r WEL}
WEL_cont_cov <- cov_combine(flow_data = WEL_flow_long,
                            spill_data = WEL_spill_long,
                            temp_data = WEL_temp_long)

WEL_cont_cov <- cov_data_availability(WEL_cont_cov)

# data availability as a table
table(WEL_cont_cov$data_available)
```

#### Data by day of year
```{r}
cov_data_availability_2(cont_cov = WEL_cont_cov)
# visualize data as plots
cov_data_viz(cont_cov = WEL_cont_cov)
```


#### Regressions and VIF
``` {r}
# regressions and vif
cov_regressions(cont_cov = WEL_cont_cov)
```
Variance inflation factors (Spill, Temperature, and Flow):
  
```{r WEL_vif}
cov_vif(cont_cov = WEL_cont_cov)
```


### Ice Harbor Dam

#### Data availability
Availability of spill/temperature/flow data, by number of days (2004-present):
```{r ICH}
ICH_cont_cov <- cov_combine(flow_data = ICH_flow_long,
                            spill_data = ICH_spill_long,
                            temp_data = ICH_temp_long)

ICH_cont_cov <- cov_data_availability(ICH_cont_cov)

# data availability as a table
table(ICH_cont_cov$data_available)
```

#### Data by day of year
```{r}
cov_data_availability_2(cont_cov = ICH_cont_cov)
# visualize data as plots
cov_data_viz(cont_cov = ICH_cont_cov)
```


#### Regressions and VIF
``` {r}
# regressions and vif
cov_regressions(cont_cov = ICH_cont_cov)
```
Variance inflation factors (Spill, Temperature, and Flow):
  
```{r ICH_vif}
cov_vif(cont_cov = ICH_cont_cov)
```



### Lower Monumental Dam

#### Data availability
Availability of spill/temperature/flow data, by number of days (2004-present):
```{r LMO}
LMO_cont_cov <- cov_combine(flow_data = LMO_flow_long,
                            spill_data = LMO_spill_long,
                            temp_data = LMO_temp_long)

LMO_cont_cov <- cov_data_availability(LMO_cont_cov)

# data availability as a table
table(LMO_cont_cov$data_available)
```

#### Data by day of year
```{r}
cov_data_availability_2(cont_cov = LMO_cont_cov)
# visualize data as plots
cov_data_viz(cont_cov = LMO_cont_cov)
```


#### Regressions and VIF
``` {r}
# regressions and vif
cov_regressions(cont_cov = LMO_cont_cov)
```
Variance inflation factors (Spill, Temperature, and Flow):
  
```{r LMO_vif}
cov_vif(cont_cov = LMO_cont_cov)
```


### Little Goose Dam

#### Data availability
Availability of spill/temperature/flow data, by number of days (2004-present):
```{r LGO}
LGO_cont_cov <- cov_combine(flow_data = LGO_flow_long,
                            spill_data = LGO_spill_long,
                            temp_data = LGO_temp_long)

LGO_cont_cov <- cov_data_availability(LGO_cont_cov)

# data availability as a table
table(LGO_cont_cov$data_available)
```

#### Data by day of year
```{r}
cov_data_availability_2(cont_cov = LGO_cont_cov)
# visualize data as plots
cov_data_viz(cont_cov = LGO_cont_cov)
```


#### Regressions and VIF
``` {r}
# regressions and vif
cov_regressions(cont_cov = LGO_cont_cov)
```
Variance inflation factors (Spill, Temperature, and Flow):
  
```{r LGO_vif}
cov_vif(cont_cov = LGO_cont_cov)
```



### Lower Granite Dam

#### Data availability
Availability of spill/temperature/flow data, by number of days (2004-present):
```{r LGR}
LGR_cont_cov <- cov_combine(flow_data = LGR_flow_long,
                            spill_data = LGR_spill_long,
                            temp_data = LGR_temp_long)

LGR_cont_cov <- cov_data_availability(LGR_cont_cov)

# data availability as a table
table(LGR_cont_cov$data_available)
```

#### Data by day of year
```{r}
cov_data_availability_2(cont_cov = LGR_cont_cov)
# visualize data as plots
cov_data_viz(cont_cov = LGR_cont_cov)
```


#### Regressions and VIF
``` {r}
# regressions and vif
cov_regressions(cont_cov = LGR_cont_cov)
```
Variance inflation factors (Spill, Temperature, and Flow):
  
```{r LGR_vif}
cov_vif(cont_cov = LGR_cont_cov)
```



## State timing

```{r state_timing}
# reformat data to add columns for month/day/day of year
ASC %>% 
  mutate(date_time = ymd_hms(date_time)) %>% 
  mutate(day = yday(date_time)) %>% 
  mutate(month = month(date_time)) %>% 
  mutate(year = year(date_time)) -> ASC

dam_pathways <- c("BON (adult)", "MCN (adult)", "PRA (adult)", "RIS (adult)",
                                      "RRE (adult)", "WEL (adult)", "ICH (adult)", "LGR (adult)")

trib_pathways <- unique(ASC$pathway)[grep("upstream|sites|RM", unique(ASC$pathway))]

# sort by tributary
ASO_pathways <- trib_pathways[grep("ASO", trib_pathways)]
CLE_pathways <- trib_pathways[grep("CLE", trib_pathways)]
DES_pathways <- trib_pathways[grep("DES", trib_pathways)]
ENT_pathways <- trib_pathways[grep("ENT", trib_pathways)]
FIF_pathways <- trib_pathways[grep("FIF", trib_pathways)]
GRRO_pathways <- trib_pathways[grep("GRRO", trib_pathways)]
HOOD_pathways <- trib_pathways[grep("HOOD", trib_pathways)]
IMN_pathways <- trib_pathways[grep("IMN", trib_pathways)]
JDR_pathways <- trib_pathways[grep("JDR", trib_pathways)]
MET_pathways <- trib_pathways[grep("MET", trib_pathways)]
OKA_pathways <- trib_pathways[grep("OKA", trib_pathways)]
SAL_pathways <- trib_pathways[grep("SAL", trib_pathways)]
TUC_pathways <- trib_pathways[grep("TUC", trib_pathways)]
UMA_pathways <- trib_pathways[grep("UMA", trib_pathways)]
WAWA_pathways <- trib_pathways[grep("WAWA", trib_pathways)]
WEN_pathways <- trib_pathways[grep("WEN", trib_pathways)]
YAK_pathways <- trib_pathways[grep("YAK", trib_pathways)]

known_fallback_pathways <- unique(ASC$pathway)[grep("fallback", unique(ASC$pathway))]

# setdiff(unique(ASC$pathway), c(trib_pathways, dam_pathways))
```

### Dam ascents
```{r dam_ascent_timing}
# BON
ggplot(subset(ASC, pathway == "BON (adult)"), aes(x = day)) +
  geom_histogram() +
  ggtitle("BON (adult)")

# MCN
ggplot(subset(ASC, pathway == "MCN (adult)"), aes(x = day)) +
  geom_histogram() +
  ggtitle("MCN (adult)")

# PRA
ggplot(subset(ASC, pathway == "PRA (adult)"), aes(x = day)) +
  geom_histogram() +
  ggtitle("PRA (adult)")

# RIS
ggplot(subset(ASC, pathway == "RIS (adult)"), aes(x = day)) +
  geom_histogram() +
  ggtitle("RIS (adult)")

# RRE
ggplot(subset(ASC, pathway == "RRE (adult)"), aes(x = day)) +
  geom_histogram() +
  ggtitle("WEL (adult)")

# WEL
ggplot(subset(ASC, pathway == "WEL (adult)"), aes(x = day)) +
  geom_histogram() + 
  ggtitle("WEL (adult)")

# ICH
ggplot(subset(ASC, pathway == "ICH (adult)"), aes(x = day)) +
  geom_histogram() +
  ggtitle("ICH (adult)")

# LGR
ggplot(subset(ASC, pathway == "LGR (adult)"), aes(x = day)) +
  geom_histogram() +
  ggtitle("LGR (adult)")
```

### Tributary entries
```{r trib_entry_timing}
# for this - let's look at entries to tributaries, not movement within
ASC %>% 
  group_by(tag_code) %>% 
  filter(!(pathway %in% trib_pathways & lag(pathway %in% trib_pathways))) -> ASC

# plot them all!
ggplot(subset(ASC, pathway %in% ASO_pathways), aes(x = day, fill = natal_origin)) +
  geom_histogram() +
  scale_fill_tableau(palette = "Tableau 20")+
  ggtitle("Asotin Creek")

ggplot(subset(ASC, pathway %in% CLE_pathways), aes(x = day, fill = natal_origin)) +
  geom_histogram() + scale_fill_tableau(palette = "Tableau 20")+
  ggtitle("Clearwater River")

ggplot(subset(ASC, pathway %in% DES_pathways), aes(x = day, fill = natal_origin)) +
  geom_histogram() + scale_fill_tableau(palette = "Tableau 20")+
  ggtitle("Deschutes River")

ggplot(subset(ASC, pathway %in% ENT_pathways), aes(x = day, fill = natal_origin)) +
  geom_histogram() + scale_fill_tableau(palette = "Tableau 20")+
  ggtitle("Entiat River")

ggplot(subset(ASC, pathway %in% FIF_pathways), aes(x = day, fill = natal_origin)) +
  geom_histogram() + scale_fill_tableau(palette = "Tableau 20")+
  ggtitle("Fifteenmile Creek")

ggplot(subset(ASC, pathway %in% GRRO_pathways), aes(x = day, fill = natal_origin)) +
  geom_histogram() + scale_fill_tableau(palette = "Tableau 20")+
  ggtitle("Grande Ronde River")

ggplot(subset(ASC, pathway %in% HOOD_pathways), aes(x = day, fill = natal_origin)) +
  geom_histogram() + scale_fill_tableau(palette = "Tableau 20")+
  ggtitle("Hood River")

ggplot(subset(ASC, pathway %in% IMN_pathways), aes(x = day, fill = natal_origin)) +
  geom_histogram() + scale_fill_tableau(palette = "Tableau 20")+
  ggtitle("Imnaha River")

ggplot(subset(ASC, pathway %in% JDR_pathways), aes(x = day, fill = natal_origin)) +
  geom_histogram() + scale_fill_tableau(palette = "Tableau 20")+
  ggtitle("John Day River")

ggplot(subset(ASC, pathway %in% MET_pathways), aes(x = day, fill = natal_origin)) +
  geom_histogram() + scale_fill_tableau(palette = "Tableau 20")+
  ggtitle("Methow River")

ggplot(subset(ASC, pathway %in% OKA_pathways), aes(x = day, fill = natal_origin)) +
  geom_histogram() + scale_fill_tableau(palette = "Tableau 20")+
  ggtitle("Okanogan River")

ggplot(subset(ASC, pathway %in% SAL_pathways), aes(x = day, fill = natal_origin)) +
  geom_histogram() + scale_fill_tableau(palette = "Tableau 20")+
  ggtitle("Salmon River")

ggplot(subset(ASC, pathway %in% TUC_pathways), aes(x = day, fill = natal_origin)) +
  geom_histogram() + scale_fill_tableau(palette = "Tableau 20")+
  ggtitle("Tucannon River")

ggplot(subset(ASC, pathway %in% UMA_pathways), aes(x = day, fill = natal_origin)) +
  geom_histogram() + scale_fill_tableau(palette = "Tableau 20")+
  ggtitle("Umatilla River")

ggplot(subset(ASC, pathway %in% WAWA_pathways), aes(x = day, fill = natal_origin)) +
  geom_histogram() + scale_fill_tableau(palette = "Tableau 20")+
  ggtitle("Walla Walla River")

ggplot(subset(ASC, pathway %in% WEN_pathways), aes(x = day, fill = natal_origin)) +
  geom_histogram() + scale_fill_tableau(palette = "Tableau 20")+
  ggtitle("Wenatchee River")

ggplot(subset(ASC, pathway %in% YAK_pathways), aes(x = day, fill = natal_origin)) +
  geom_histogram() + scale_fill_tableau(palette = "Tableau 20")+
  ggtitle("Yakima River")
```


### Dam fallback (timing estimated)
