==> R/25_spill_data_prep.Rmd <==
---
title: "25_spill_data_prep"
author: "Markus Min"
date: "2023-07-10"
output: html_document
---

#### Description
This Rmd will explain the rationale for how spill data was set up for inclusion in our model and also process the spill data.


```{r libraries-setup, echo = FALSE, message = FALSE, warning = FALSE}
knitr::opts_chunk$set(echo = FALSE, message = FALSE, warning = FALSE)

# load libraries
library(here)
library(janitor)
library(tidyverse)
library(lubridate)
library(car)
library(ggthemes)
library(viridis)
library(mixtools)
```



```{r load_reformat_data}
# Function 1: reformat covariate data

# Spill is in kcfs (1000 cu ft/s), but the scale of this is still too high. To get it approx
# on the same scale as temp (which is z-scored), divide by 100. This means that the units
# now are in 100s of kcfs
spill_reformat <- function(input_data, variable_name){
  
  # need to deal with leap years
  leap_years <- seq(1960, 2024, by = 4)
  
  input_data %>% 
    pivot_longer(., cols = colnames(.)[2:ncol(.)]) %>% 
    dplyr::rename(year = name,
                  !!quo_name(variable_name) := value) %>% 
    mutate(year = gsub("X","",year)) %>% 
    mutate(spill = spill/100) %>% 
    mutate(year = as.numeric(year), day = as.numeric(day)) %>% 
    # drop day 366 in non-leap years
    dplyr::filter(!(day == 366 & !(year %in% leap_years))) %>%
    # Add a new column for date (combine year and day)
    mutate(date = format(as.Date(day, origin = paste0(year-1, "-12-31")))) %>% 
    arrange(year, day) -> output_data
  
  return(output_data)
}

# load and reformat the spill data

# BON
BON_spill <- read.csv(here::here("covariate_data", "BON", "basin_spill_BON_allyears.csv"))[1:366,]
BON_spill_long <- spill_reformat(input_data = BON_spill, variable_name = "spill")

# ICH
ICH_spill <- read.csv(here::here("covariate_data", "ICH", "basin_spill_IHR_allyears.csv"))[1:366,]
ICH_spill_long <- spill_reformat(input_data = ICH_spill, variable_name = "spill")

# JDA
JDA_spill <- read.csv(here::here("covariate_data", "JDA", "basin_spill_JDA_allyears.csv"))[1:366,]
JDA_spill_long <- spill_reformat(input_data = JDA_spill, variable_name = "spill")

# TDA
TDA_spill <- read.csv(here::here("covariate_data", "TDA", "basin_spill_TDA_allyears.csv"))[1:366,]
TDA_spill_long <- spill_reformat(input_data = TDA_spill, variable_name = "spill")

# LGR
LGR_spill <- read.csv(here::here("covariate_data", "LGR", "basin_spill_LWG_allyears.csv"))[1:366,]
LGR_spill_long <- spill_reformat(input_data = LGR_spill, variable_name = "spill")

# LMO
LMO_spill <- read.csv(here::here("covariate_data", "LMO", "basin_spill_LMN_allyears.csv"))[1:366,]
LMO_spill_long <- spill_reformat(input_data = LMO_spill, variable_name = "spill")

# LGO
LGO_spill <- read.csv(here::here("covariate_data", "LGO", "basin_spill_LGS_allyears.csv"))[1:366,]
LGO_spill_long <- spill_reformat(input_data = LGO_spill, variable_name = "spill")

# MCN
MCN_spill <- read.csv(here::here("covariate_data", "MCN", "basin_spill_MCN_allyears.csv"))[1:366,]
MCN_spill_long <- spill_reformat(input_data = MCN_spill, variable_name = "spill")

# PRA
PRA_spill <- read.csv(here::here("covariate_data", "PRA", "basin_spill_PRD_allyears.csv"))[1:366,]
PRA_spill_long <- spill_reformat(input_data = PRA_spill, variable_name = "spill")

# RIS
RIS_spill <- read.csv(here::here("covariate_data", "RIS", "basin_spill_RIS_allyears.csv"))[1:366,]
RIS_spill_long <- spill_reformat(input_data = RIS_spill, variable_name = "spill")

# RRE
RRE_spill <- read.csv(here::here("covariate_data", "RRE", "basin_spill_RRH_allyears.csv"))[1:366,]
RRE_spill_long <- spill_reformat(input_data = RRE_spill, variable_name = "spill")

# WEL
WEL_spill <- read.csv(here::here("covariate_data", "WEL", "basin_spill_WEL_allyears.csv"))[1:366,]
WEL_spill_long <- spill_reformat(input_data = WEL_spill, variable_name = "spill")

# so, some gaps in WEL spill data, on Jan 1 and 2 2020, and in December 2021.
# I think it's safe to assume that these are zero, especially looking at typical spill during these months
# let's look at that data
ggplot(filter(WEL_spill_long, year %in% c(2020, 2021)), aes(x = ymd(date), y = spill)) + geom_point()
# yeah, most likely zeros. will interpolate as such

WEL_spill_long %>% 
  mutate(spill = ifelse(is.na(spill) & year < 2023, 0, spill)) -> WEL_spill_long

```

### data exploration
```{r}
spill_by_year_plot <- function(data, dam_name){
  ggplot(subset(data, year >= 2005), aes(x = day, y = spill, color = year)) +
  # scale_color_tableau(palette = "Tableau 20") +
  scale_color_viridis() +
  geom_point() +
  ggtitle(dam_name)

}

spill_by_year_plot(data = BON_spill_long, dam_name = "Bonneville")
spill_by_year_plot(data = MCN_spill_long, dam_name = "McNary")
spill_by_year_plot(data = PRA_spill_long, dam_name = "Priest Rapids")
spill_by_year_plot(data = RIS_spill_long, dam_name = "Rock Island")
spill_by_year_plot(data = RRE_spill_long, dam_name = "Rocky Reach")
spill_by_year_plot(data = WEL_spill_long, dam_name = "Wells")
spill_by_year_plot(data = ICH_spill_long, dam_name = "Ice Harbor")
spill_by_year_plot(data = LGR_spill_long, dam_name = "Lower Granite")

```


# Spill data processing for model inclusion

We are interested in how spill affects fallback rates, both post-overshoot and en-route. However, given how these processes are fundamentally different, we will process spill data in different ways based on our hypotheses of how spill interact with these processes.



## Two hypotheses

#### En-route fallback

En-route fallback occurs when fish fall back prior to reaching their natal tributaries. This type of fallback **goes against the desired direction of movement for all fish**, as their goal is to continue traveling upstream. Our hypothesis surrounding en-route fallback is that **it is involuntary; i.e., that it occurs as a result of river conditions (spill, flow) and/or individual condition**. As such, **our hypothesis for the relationship between spill and en-route fallback is that it is related to the volume of spill, as this would serve to push fish back downstream via the spillway**. The approach we will use for this spill covariate is the same as was used for temperature, i.e., the use of windows and taking the average spill during that window.


#### Post-overshoot fallback

Post-overshoot fallback occurs when fish fall back after overshooting natal tributaries (having ascended the dam directly upstream of their natal tributary). This type of fallback **is a desirable outcome for all fish**, as this movement helps them get closer to their natal tributary.  Our hypothesis surrounding post-overshoot fallback is that **it is voluntary; i.e., that it occurs when fish make the decision to want to go back downstream and towards their natal tributaries**. As such, **our hypothesis for the relationship between spill and post-overshoot fallback is that it is related to the presence of spill, as the presence of spill allows fish to find a safe route back downstream**. This is based in large part on the work of Richins & Skalski (2018) and Wertheimer (2007), who found that even small amounts of surface flow were effective for routing steelhead kelts away from turbines, indicating that the availability of surface passage options may be more important than the volume of flow. The approach we will use for this covariate is to find the number of days with any amount of spill in the months of January, February, and March, which is when post-overshoot fallback is predicted to occur (after overwintering).


## En-route data processing

### Calculate windows - note that this is exactly the same code as is in script 24, which calculates windows for temperature

### calculate windows

```{r}
# so, I think that the most up to date files are in the same folders as the model runs
snake_adults_states_complete <- read.csv(here::here("stan_actual", "deteff_ESU_models", "500iter_3chain_runs", "snake", "snake_adults_states_complete.csv"), row.names = 1)
lowcol_adults_states_complete <- read.csv(here::here("stan_actual", "deteff_ESU_models", "lower_columbia", "lower_columbia_adults_states_complete.csv"), row.names = 1)
midcol_adults_states_complete <- read.csv(here::here("stan_actual", "deteff_ESU_models", "500iter_3chain_runs", "middle_columbia", "middle_columbia_adults_states_complete.csv"), row.names = 1)
uppcol_adults_states_complete <- read.csv(here::here("stan_actual", "deteff_ESU_models", "500iter_3chain_runs", "upper_columbia", "upper_columbia_adults_states_complete.csv"), row.names = 1)

# combine them all
snake_adults_states_complete %>% 
  bind_rows(., midcol_adults_states_complete) %>% 
  bind_rows(., uppcol_adults_states_complete) %>% 
  bind_rows(., lowcol_adults_states_complete) -> ASC

# now add tag code metadata, for natal origins
origin_numeric <- data.frame(natal_origin = c("Asotin_Creek", 
                                        "Clearwater_River",
                                        "Deschutes_River", 
                                        "Entiat_River", 
                                        "Fifteenmile_Creek", 
                                        "Grande_Ronde_River", 
                                        "Hood_River",
                                        "Imnaha_River",
                                        "John_Day_River", 
                                        "Methow_River", 
                                        "Okanogan_River", 
                                        "Salmon_River", 
                                        "Tucannon_River", 
                                        "Umatilla_River",
                                        "Walla_Walla_River",
                                        "Wenatchee_River", 
                                        "Yakima_River"),
                             natal_origin_numeric = seq(1,17,1))

origin_rear_actual <- read.csv(here::here("stan_actual", "origin_rear_actual.csv"), row.names = 1)
origin_rear_actual %>% 
  dplyr::rename(natal_origin_numeric = natal_origin) %>% 
  left_join(origin_numeric, by = "natal_origin_numeric") %>% 
  dplyr::select(tag_code_2, natal_origin) -> tag_code_origins

ASC %>% 
  left_join(., tag_code_origins, by = "tag_code_2") -> ASC

# also note which ESU they're from, for plotting
ESU_origins <- data.frame(natal_origin = unique(ASC$natal_origin),
                          ESU = c(rep("Snake", 6),
                                  rep("Middle Columbia", 6),
                                  rep("Upper Columbia", 4),
                                  "Lower Columbia"))

ASC %>% 
  left_join(., ESU_origins, by = "natal_origin") -> ASC

# a function that takes just one state, and finds the average amount of time to move out of that state

residence_time <- function(residence_state){
  
  # don't keep any with implicit time interpolated
  ASC %>% 
    filter(state == residence_state & tag_code_2 == lead(tag_code_2) & pathway != "implicit" & lead(pathway) != "implicit" |
             lag(state) == residence_state & tag_code_2 == lag(tag_code_2) & pathway != "implicit" & lag(pathway) != "implicit") %>% 
    mutate(date_time = ymd_hms(date_time)) -> one_state_df
  
  # make a data frame to record all of the transitions
  n_transitions <- nrow(one_state_df)/2
  passage_df <- data.frame(tag_code_2 = one_state_df$tag_code_2[seq(1,(nrow(one_state_df)-1),2)],
                           ESU = one_state_df$ESU[seq(1,(nrow(one_state_df)-1),2)],
                           natal_origin = one_state_df$natal_origin[seq(1,(nrow(one_state_df)-1),2)],
                           passage_time = NA)
  
  
  for (i in 1:nrow(passage_df)){
    passage_df$passage_time[i] <- one_state_df$date_time[(i*2)] - one_state_df$date_time[(i*2-1)]
    
  }
  
  return(passage_df)
  
}

main_mainstem_states <- c(
  "mainstem, BON to MCN",
  "mainstem, MCN to ICH or PRA",
  "mainstem, PRA to RIS",
  "mainstem, RIS to RRE",
  "mainstem, RRE to WEL",
  "mainstem, ICH to LGR",
  "mainstem, upstream of WEL",
  "mainstem, upstream of LGR")

list_quantiles <- list()

for (i in 1:length(main_mainstem_states)) {
  residence_df <- residence_time(residence_state = main_mainstem_states[i])
  
  # Look into windows - show: 1) median of the whole distribution, 
  # and what cutoff you'd need to capture 50%, 80%, and 95% of fish residence time
  quantiles <- quantile(residence_df$passage_time, probs = c(0.25, 0.5, 0.75, 0.8, 0.95))
  list_quantiles[[i]] <- quantiles

  
}

# quantiles isn't really going to work for upstream LGR and WEL, need a mixture model for those

# Upstream of Lower Granite
LGR_residence_df <- residence_time(residence_state = "mainstem, upstream of LGR")
LGR_mix <- normalmixEM(LGR_residence_df$passage_time, k = 2)

# most fish are longer residence time (81% vs. shorter residence time)
LGR_mix$lambda
# shorter residence time is 24 days, longer is 168 days
LGR_mix$mu

# show the fit
plot(LGR_mix, density=TRUE, cex.axis=1.4, cex.lab=1.4, cex.main=1.8,
      main2="Residence time, upstream of LGR", xlab2="Days")


# Upstream of Wells
WEL_residence_df <- residence_time(residence_state = "mainstem, upstream of WEL")
WEL_mix <- normalmixEM(WEL_residence_df$passage_time, k = 2)

# 51%/49% for short vs. long residence time at Wells, much different than at LGR
WEL_mix$lambda
# shorter residence time is 21 days, longer is 187 days
WEL_mix$mu

# show the fit
plot(WEL_mix, density=TRUE, cex.axis=1.4, cex.lab=1.4, cex.main=1.8,
      main2="Residence time, upstream of WEL", xlab2="Days")


# make an actual table
quantiles_table <- data.frame(reach = main_mainstem_states,
                              t25 = c(list_quantiles[[1]][1],
                                       list_quantiles[[2]][1],
                                       list_quantiles[[3]][1],
                                       list_quantiles[[4]][1],
                                       list_quantiles[[5]][1],
                                       list_quantiles[[6]][1],
                                      list_quantiles[[7]][1],
                                      list_quantiles[[8]][1]),
                              t50 = c(list_quantiles[[1]][2],
                                       list_quantiles[[2]][2],
                                       list_quantiles[[3]][2],
                                       list_quantiles[[4]][2],
                                       list_quantiles[[5]][2],
                                       list_quantiles[[6]][2],
                                      list_quantiles[[7]][2],
                                      list_quantiles[[8]][2]),
                              t75 = c(list_quantiles[[1]][3],
                                       list_quantiles[[2]][3],
                                       list_quantiles[[3]][3],
                                       list_quantiles[[4]][3],
                                       list_quantiles[[5]][3],
                                       list_quantiles[[6]][3],
                                      list_quantiles[[7]][3],
                                      list_quantiles[[8]][3]),
                              t80 = c(list_quantiles[[1]][4],
                                       list_quantiles[[2]][4],
                                       list_quantiles[[3]][4],
                                       list_quantiles[[4]][4],
                                       list_quantiles[[5]][4],
                                       list_quantiles[[6]][4],
                                      list_quantiles[[7]][4],
                                      list_quantiles[[8]][4]),
                              t95 = c(list_quantiles[[1]][5],
                                       list_quantiles[[2]][5],
                                       list_quantiles[[3]][5],
                                       list_quantiles[[4]][5],
                                       list_quantiles[[5]][5],
                                       list_quantiles[[6]][5],
                                      list_quantiles[[7]][5],
                                      list_quantiles[[8]][5]))

# write.csv(quantiles_table, here::here("covariate_data", "reach_quantiles.csv"))
```

### export spill data using window approach

```{r}
# Re-export, using reach-specific residence time windows
# quantiles <- read.csv(here::here("covariate_data", "reach_quantiles.csv"), row.names = 1)

# round the median to nearest whole day
quantiles_table %>% 
  mutate(median = round(t50)) -> quantiles_table

# match them using outflow

# use a function
# optional argument if you want to start window at a different time
# also update to make it flexible at the end of the time series to shorten window at the end of the time series (but this shouldn't really matter, we're not really seeing any movements in December 2022); but it'll just take the spill at that day
window_spill <- function(spill_data, start_window_days = 0, end_window_days){
  spill_data %>% 
    mutate(date = ymd(date)) %>% 
    filter(date >= ymd("2005-06-01") & date <= ymd("2022-12-31")) %>% 
    dplyr::select(date, spill) %>% 
    relocate(date) -> spill_data
  
  spill_data %>% 
  mutate(window_spill = NA) -> spill_data

# loop to get all windows
for (i in 1:nrow(spill_data)){
  if (i > nrow(spill_data)-start_window_days) {
    spill_data$window_spill[i] <- spill_data$spill[i]
    
  } else {
    spill_data$window_spill[i] <- mean(subset(spill_data, date >= spill_data$date[i] + days(x = start_window_days) & date <= spill_data$date[i] + days(x = end_window_days))$spill)    
    
  }
  

}
  
  return(spill_data)
}

# Bonneville Dam, for mainstem, BON to MCN reach
BON_window_spill <- window_spill(spill_data = BON_spill_long, end_window_days = subset(quantiles_table, reach == "mainstem, BON to MCN")$median)

# McNary Dam, for mainstem, MCN to ICH or PRA reach
MCN_window_spill <- window_spill(spill_data = MCN_spill_long, end_window_days = subset(quantiles_table, reach == "mainstem, MCN to ICH or PRA")$median)

# PRA Dam, for PRA to RIS reach
PRA_window_spill <- window_spill(spill_data = PRA_spill_long, end_window_days = subset(quantiles_table, reach == "mainstem, PRA to RIS")$median)

# RIS Dam, for RIS to RRE reach
RIS_window_spill <- window_spill(spill_data = RIS_spill_long, end_window_days = subset(quantiles_table, reach == "mainstem, RIS to RRE")$median)

# RRE Dam, for RRE to WEL reach
RRE_window_spill <- window_spill(spill_data = RRE_spill_long, end_window_days = subset(quantiles_table, reach == "mainstem, RRE to WEL")$median)

# Ice Harbor Dam, for ICH to LGR reach
ICH_window_spill <- window_spill(spill_data = ICH_spill_long, end_window_days = subset(quantiles_table, reach == "mainstem, ICH to LGR")$median)

# Wells Dam, for upstream of WEL reach - quick fish
WEL_quick_index <- which.min(WEL_mix$mu)
WEL_slow_index <- which.max(WEL_mix$mu)


WEL_quick_window_spill <- window_spill(spill_data = WEL_spill_long, 
                                     start_window_days = round(WEL_mix$mu[WEL_quick_index] - WEL_mix$sigma[WEL_quick_index]), 
                                     end_window_days = round(WEL_mix$mu[WEL_quick_index] + WEL_mix$sigma[WEL_quick_index]))

# Wells Dam, for upstream of WEL reach - slow fish
WEL_slow_window_spill <- window_spill(spill_data = WEL_spill_long, 
                                     start_window_days = round(WEL_mix$mu[WEL_slow_index] - WEL_mix$sigma[WEL_slow_index]), 
                                     end_window_days = round(WEL_mix$mu[WEL_slow_index] + WEL_mix$sigma[WEL_slow_index]))

# Lower Granite Dam, for upstream of LGR reach - quick fish
LGR_quick_index <- which.min(LGR_mix$mu)
LGR_slow_index <- which.max(LGR_mix$mu)

LGR_quick_window_spill <- window_spill(spill_data = LGR_spill_long, 
                                     start_window_days = round(LGR_mix$mu[LGR_quick_index] - LGR_mix$sigma[LGR_quick_index]), 
                                     end_window_days = round(LGR_mix$mu[LGR_quick_index] + LGR_mix$sigma[LGR_quick_index]))

# Lower Granite Dam, for upstream of LGR reach - slow fish
LGR_slow_window_spill <- window_spill(spill_data = LGR_spill_long, 
                                     start_window_days = round(LGR_mix$mu[LGR_slow_index] - LGR_mix$sigma[LGR_slow_index]), 
                                     end_window_days = round(LGR_mix$mu[LGR_slow_index] + LGR_mix$sigma[LGR_slow_index]))

# export all of these
write.csv(BON_window_spill, here::here("covariate_data", "for_model", "spill", "BON_window_spill.csv"), row.names = FALSE)
write.csv(MCN_window_spill, here::here("covariate_data", "for_model", "spill", "MCN_window_spill.csv"), row.names = FALSE)
write.csv(PRA_window_spill, here::here("covariate_data", "for_model", "spill", "PRA_window_spill.csv"), row.names = FALSE)
write.csv(RIS_window_spill, here::here("covariate_data", "for_model", "spill", "RIS_window_spill.csv"), row.names = FALSE)
write.csv(RRE_window_spill, here::here("covariate_data", "for_model", "spill", "RRE_window_spill.csv"), row.names = FALSE)
write.csv(ICH_window_spill, here::here("covariate_data", "for_model", "spill", "ICH_window_spill.csv"), row.names = FALSE)
write.csv(WEL_quick_window_spill, here::here("covariate_data", "for_model", "spill", "WEL_quick_window_spill.csv"), row.names = FALSE)
write.csv(WEL_slow_window_spill, here::here("covariate_data", "for_model", "spill", "WEL_slow_window_spill.csv"), row.names = FALSE)
write.csv(LGR_quick_window_spill, here::here("covariate_data", "for_model", "spill", "LGR_quick_window_spill.csv"), row.names = FALSE)
write.csv(LGR_slow_window_spill, here::here("covariate_data", "for_model", "spill", "LGR_slow_window_spill.csv"), row.names = FALSE)

```

#### export in the format needed for stan
```{r}
dplyr::rename(dplyr::select(BON_window_spill, date, window_spill), BON = window_spill) %>% 
  left_join(dplyr::rename(dplyr::select(MCN_window_spill, date, window_spill), MCN = window_spill), by = "date") %>% 
  left_join(dplyr::rename(dplyr::select(PRA_window_spill, date, window_spill), PRA = window_spill), by = "date") %>% 
  left_join(dplyr::rename(dplyr::select(RIS_window_spill, date, window_spill), RIS = window_spill), by = "date") %>% 
  left_join(dplyr::rename(dplyr::select(RRE_window_spill, date, window_spill), RRE = window_spill), by = "date") %>% 
  left_join(dplyr::rename(dplyr::select(WEL_quick_window_spill, date, window_spill), WEL_quick = window_spill), by = "date") %>% 
  left_join(dplyr::rename(dplyr::select(ICH_window_spill, date, window_spill), ICH = window_spill), by = "date") %>% 
  left_join(dplyr::rename(dplyr::select(LGR_quick_window_spill, date, window_spill), LGR_quick = window_spill), by = "date") %>% 
  left_join(dplyr::rename(dplyr::select(WEL_slow_window_spill, date, window_spill), WEL_slow = window_spill), by = "date") %>%
  left_join(dplyr::rename(dplyr::select(LGR_slow_window_spill, date, window_spill), LGR_slow = window_spill), by = "date") %>%
  mutate(index = row_number()) %>% 
  relocate(index) %>% 
  dplyr::select(-date) %>% 
  column_to_rownames("index") -> window_spill

# add an empty column for the first state (that doesn't get a spill effect)
# make sure that the order is the same is the order of states
window_spill %>% 
  mutate(MOUTH = 0) %>% 
  relocate(MOUTH) -> window_spill
  
write.csv(window_spill, here::here("covariate_data", "for_model", "window_spill_for_stan.csv"))
```



## Post-overshoot processing

Here, we will also divide number of days by 100, to rescale

### Export January, February, March, and April days of spill as covariate
```{r}
april_spill_days <- function(spill_data){
  spill_data %>% 
    # keep only 2005-2022
    subset(year >= 2005 & year <= 2022) %>% 
    mutate(date = ymd(date)) %>% 
    mutate(month = month(date)) %>% 
    mutate(any_spill = ifelse(spill > 0, "yes", "no")) %>% 
    group_by(year, month) %>% 
    count(any_spill == "yes") %>% 
    dplyr::rename(positive_spill = `any_spill == "yes"`) %>% 
    subset(positive_spill == TRUE & month == 4) %>% 
    dplyr::rename(april_spill_days = n) %>% 
    mutate(april_spill_days = april_spill_days/100) %>% 
    ungroup() %>% 
    complete(year = seq(2005,2022,1), fill = list(april_spill_days = 0)) %>% 
    dplyr::select(year, april_spill_days) -> april_spill_days
  
  return(april_spill_days)
}

march_spill_days <- function(spill_data){
  spill_data %>% 
    # keep only 2005-2022
    subset(year >= 2005 & year <= 2022) %>% 
    mutate(date = ymd(date)) %>% 
    mutate(month = month(date)) %>% 
    mutate(any_spill = ifelse(spill > 0, "yes", "no")) %>% 
    group_by(year, month) %>% 
    count(any_spill == "yes") %>% 
    dplyr::rename(positive_spill = `any_spill == "yes"`) %>% 
    subset(positive_spill == TRUE & month == 3) %>% 
    dplyr::rename(march_spill_days = n) %>% 
    mutate(march_spill_days = march_spill_days/100) %>% 
    ungroup() %>% 
    complete(year = seq(2005,2022,1), fill = list(march_spill_days = 0)) %>% 
    dplyr::select(year, march_spill_days) -> march_spill_days
  
  return(march_spill_days)
}

february_spill_days <- function(spill_data){
  spill_data %>% 
    # keep only 2005-2022
    subset(year >= 2005 & year <= 2022) %>% 
    mutate(date = ymd(date)) %>% 
    mutate(month = month(date)) %>% 
    mutate(any_spill = ifelse(spill > 0, "yes", "no")) %>% 
    group_by(year, month) %>% 
    count(any_spill == "yes") %>% 
    dplyr::rename(positive_spill = `any_spill == "yes"`) %>% 
    subset(positive_spill == TRUE & month == 2) %>% 
    dplyr::rename(february_spill_days = n) %>% 
    mutate(february_spill_days = february_spill_days/100) %>% 
    ungroup() %>% 
    complete(year = seq(2005,2022,1), fill = list(february_spill_days = 0)) %>% 
    dplyr::select(year, february_spill_days) -> february_spill_days
  
  return(february_spill_days)
}

january_spill_days <- function(spill_data){
  spill_data %>% 
    # keep only 2005-2022
    subset(year >= 2005 & year <= 2022) %>% 
    mutate(date = ymd(date)) %>% 
    mutate(month = month(date)) %>% 
    mutate(any_spill = ifelse(spill > 0, "yes", "no")) %>% 
    group_by(year, month) %>% 
    count(any_spill == "yes") %>% 
    dplyr::rename(positive_spill = `any_spill == "yes"`) %>% 
    subset(positive_spill == TRUE & month == 1) %>% 
    dplyr::rename(january_spill_days = n) %>% 
    mutate(january_spill_days = january_spill_days/100) %>% 
    ungroup() %>% 
    complete(year = seq(2005,2022,1), fill = list(january_spill_days = 0)) %>% 
    dplyr::select(year, january_spill_days) -> january_spill_days
  
  return(january_spill_days)
}

BON_april_spill <- april_spill_days(spill_data = BON_spill_long)
MCN_april_spill <- april_spill_days(spill_data = MCN_spill_long)
PRA_april_spill <- april_spill_days(spill_data = PRA_spill_long)
RIS_april_spill <- april_spill_days(spill_data = RIS_spill_long)
RRE_april_spill <- april_spill_days(spill_data = RRE_spill_long)
WEL_april_spill <- april_spill_days(spill_data = WEL_spill_long)
ICH_april_spill <- april_spill_days(spill_data = ICH_spill_long)
LGR_april_spill <- april_spill_days(spill_data = LGR_spill_long)


BON_march_spill <- march_spill_days(spill_data = BON_spill_long)
MCN_march_spill <- march_spill_days(spill_data = MCN_spill_long)
PRA_march_spill <- march_spill_days(spill_data = PRA_spill_long)
RIS_march_spill <- march_spill_days(spill_data = RIS_spill_long)
RRE_march_spill <- march_spill_days(spill_data = RRE_spill_long)
WEL_march_spill <- march_spill_days(spill_data = WEL_spill_long)
ICH_march_spill <- march_spill_days(spill_data = ICH_spill_long)
LGR_march_spill <- march_spill_days(spill_data = LGR_spill_long)

BON_february_spill <- february_spill_days(spill_data = BON_spill_long)
MCN_february_spill <- february_spill_days(spill_data = MCN_spill_long)
PRA_february_spill <- february_spill_days(spill_data = PRA_spill_long)
RIS_february_spill <- february_spill_days(spill_data = RIS_spill_long)
RRE_february_spill <- february_spill_days(spill_data = RRE_spill_long)
WEL_february_spill <- february_spill_days(spill_data = WEL_spill_long)
ICH_february_spill <- february_spill_days(spill_data = ICH_spill_long)
LGR_february_spill <- february_spill_days(spill_data = LGR_spill_long)

BON_january_spill <- january_spill_days(spill_data = BON_spill_long)
MCN_january_spill <- january_spill_days(spill_data = MCN_spill_long)
PRA_january_spill <- january_spill_days(spill_data = PRA_spill_long)
RIS_january_spill <- january_spill_days(spill_data = RIS_spill_long)
RRE_january_spill <- january_spill_days(spill_data = RRE_spill_long)
WEL_january_spill <- january_spill_days(spill_data = WEL_spill_long)
ICH_january_spill <- january_spill_days(spill_data = ICH_spill_long)
LGR_january_spill <- january_spill_days(spill_data = LGR_spill_long)

write.csv(BON_april_spill, here::here("covariate_data", "for_model", "spill", "BON_april_spill.csv"), row.names = FALSE)
write.csv(MCN_april_spill, here::here("covariate_data", "for_model", "spill", "MCN_april_spill.csv"), row.names = FALSE)
write.csv(PRA_april_spill, here::here("covariate_data", "for_model", "spill", "PRA_april_spill.csv"), row.names = FALSE)
write.csv(RIS_april_spill, here::here("covariate_data", "for_model", "spill", "RIS_april_spill.csv"), row.names = FALSE)
write.csv(RRE_april_spill, here::here("covariate_data", "for_model", "spill", "RRE_april_spill.csv"), row.names = FALSE)
write.csv(WEL_april_spill, here::here("covariate_data", "for_model", "spill", "WEL_april_spill.csv"), row.names = FALSE)
write.csv(ICH_april_spill, here::here("covariate_data", "for_model", "spill", "ICH_april_spill.csv"), row.names = FALSE)
write.csv(LGR_april_spill, here::here("covariate_data", "for_model", "spill", "LGR_april_spill.csv"), row.names = FALSE)

write.csv(BON_march_spill, here::here("covariate_data", "for_model", "spill", "BON_march_spill.csv"), row.names = FALSE)
write.csv(MCN_march_spill, here::here("covariate_data", "for_model", "spill", "MCN_march_spill.csv"), row.names = FALSE)
write.csv(PRA_march_spill, here::here("covariate_data", "for_model", "spill", "PRA_march_spill.csv"), row.names = FALSE)
write.csv(RIS_march_spill, here::here("covariate_data", "for_model", "spill", "RIS_march_spill.csv"), row.names = FALSE)
write.csv(RRE_march_spill, here::here("covariate_data", "for_model", "spill", "RRE_march_spill.csv"), row.names = FALSE)
write.csv(WEL_march_spill, here::here("covariate_data", "for_model", "spill", "WEL_march_spill.csv"), row.names = FALSE)
write.csv(ICH_march_spill, here::here("covariate_data", "for_model", "spill", "ICH_march_spill.csv"), row.names = FALSE)
write.csv(LGR_march_spill, here::here("covariate_data", "for_model", "spill", "LGR_march_spill.csv"), row.names = FALSE)

write.csv(BON_february_spill, here::here("covariate_data", "for_model", "spill", "BON_february_spill.csv"), row.names = FALSE)
write.csv(MCN_february_spill, here::here("covariate_data", "for_model", "spill", "MCN_february_spill.csv"), row.names = FALSE)
write.csv(PRA_february_spill, here::here("covariate_data", "for_model", "spill", "PRA_february_spill.csv"), row.names = FALSE)
write.csv(RIS_february_spill, here::here("covariate_data", "for_model", "spill", "RIS_february_spill.csv"), row.names = FALSE)
write.csv(RRE_february_spill, here::here("covariate_data", "for_model", "spill", "RRE_february_spill.csv"), row.names = FALSE)
write.csv(WEL_february_spill, here::here("covariate_data", "for_model", "spill", "WEL_february_spill.csv"), row.names = FALSE)
write.csv(ICH_february_spill, here::here("covariate_data", "for_model", "spill", "ICH_february_spill.csv"), row.names = FALSE)
write.csv(LGR_february_spill, here::here("covariate_data", "for_model", "spill", "LGR_february_spill.csv"), row.names = FALSE)

write.csv(BON_january_spill, here::here("covariate_data", "for_model", "spill", "BON_january_spill.csv"), row.names = FALSE)
write.csv(MCN_january_spill, here::here("covariate_data", "for_model", "spill", "MCN_january_spill.csv"), row.names = FALSE)
write.csv(PRA_january_spill, here::here("covariate_data", "for_model", "spill", "PRA_january_spill.csv"), row.names = FALSE)
write.csv(RIS_january_spill, here::here("covariate_data", "for_model", "spill", "RIS_january_spill.csv"), row.names = FALSE)
write.csv(RRE_january_spill, here::here("covariate_data", "for_model", "spill", "RRE_january_spill.csv"), row.names = FALSE)
write.csv(WEL_january_spill, here::here("covariate_data", "for_model", "spill", "WEL_january_spill.csv"), row.names = FALSE)
write.csv(ICH_january_spill, here::here("covariate_data", "for_model", "spill", "ICH_january_spill.csv"), row.names = FALSE)
write.csv(LGR_january_spill, here::here("covariate_data", "for_model", "spill", "LGR_january_spill.csv"), row.names = FALSE)

# make a quick plot for spill availability
BON_march_spill %>% 
  left_join(., BON_april_spill, by = "year") %>% 
  left_join(., BON_february_spill, by = "year") %>% 
  left_join(., BON_january_spill, by = "year") %>% 
  pivot_longer(., cols = c("april_spill_days", "march_spill_days", "february_spill_days", "january_spill_days")) -> BON_monthly_spill

ggplot(BON_monthly_spill, aes(x = year, y = value, color = name)) +
  geom_point(position = position_dodge(width = 0.75)) +
  ggtitle("BON")

MCN_march_spill %>% 
  left_join(., MCN_april_spill, by = "year") %>% 
  left_join(., MCN_february_spill, by = "year") %>% 
  left_join(., MCN_january_spill, by = "year") %>% 
  pivot_longer(., cols = c("april_spill_days", "march_spill_days", "february_spill_days", "january_spill_days")) -> MCN_monthly_spill

ggplot(MCN_monthly_spill, aes(x = year, y = value, color = name)) +
  geom_point(position = position_dodge(width = 0.75)) +
  ggtitle("MCN")

PRA_march_spill %>% 
  left_join(., PRA_april_spill, by = "year") %>% 
  left_join(., PRA_february_spill, by = "year") %>% 
  left_join(., PRA_january_spill, by = "year") %>% 
  pivot_longer(., cols = c("april_spill_days", "march_spill_days", "february_spill_days", "january_spill_days")) -> PRA_monthly_spill

ggplot(PRA_monthly_spill, aes(x = year, y = value, color = name)) +
  geom_point(position = position_dodge(width = 0.75)) +
  ggtitle("PRA")

RIS_march_spill %>% 
  left_join(., RIS_april_spill, by = "year") %>% 
  left_join(., RIS_february_spill, by = "year") %>% 
  left_join(., RIS_january_spill, by = "year") %>% 
  pivot_longer(., cols = c("april_spill_days", "march_spill_days", "february_spill_days", "january_spill_days")) -> RIS_monthly_spill

ggplot(RIS_monthly_spill, aes(x = year, y = value, color = name)) +
  geom_point(position = position_dodge(width = 0.75)) +
  ggtitle("RIS")

RRE_march_spill %>% 
  left_join(., RRE_april_spill, by = "year") %>% 
  left_join(., RRE_february_spill, by = "year") %>% 
  left_join(., RRE_january_spill, by = "year") %>% 
  pivot_longer(., cols = c("april_spill_days", "march_spill_days", "february_spill_days", "january_spill_days")) -> RRE_monthly_spill

ggplot(RRE_monthly_spill, aes(x = year, y = value, color = name)) +
  geom_point(position = position_dodge(width = 0.75)) +
  ggtitle("RRE")

WEL_march_spill %>% 
  left_join(., WEL_april_spill, by = "year") %>% 
  left_join(., WEL_february_spill, by = "year") %>% 
  left_join(., WEL_january_spill, by = "year") %>% 
  pivot_longer(., cols = c("april_spill_days", "march_spill_days", "february_spill_days", "january_spill_days")) -> WEL_monthly_spill

ggplot(WEL_monthly_spill, aes(x = year, y = value, color = name)) +
  geom_point(position = position_dodge(width = 0.75)) +
  ggtitle("WEL")

ICH_march_spill %>% 
  left_join(., ICH_april_spill, by = "year") %>% 
  left_join(., ICH_february_spill, by = "year") %>% 
  left_join(., ICH_january_spill, by = "year") %>% 
  pivot_longer(., cols = c("april_spill_days", "march_spill_days", "february_spill_days", "january_spill_days")) -> ICH_monthly_spill

ggplot(ICH_monthly_spill, aes(x = year, y = value, color = name)) +
  geom_point(position = position_dodge(width = 0.75)) +
  ggtitle("ICH")

LGR_march_spill %>% 
  left_join(., LGR_april_spill, by = "year") %>% 
  left_join(., LGR_february_spill, by = "year") %>% 
  left_join(., LGR_january_spill, by = "year") %>% 
  pivot_longer(., cols = c("april_spill_days", "march_spill_days", "february_spill_days", "january_spill_days")) -> LGR_monthly_spill

ggplot(LGR_monthly_spill, aes(x = year, y = value, color = name)) +
  geom_point(position = position_dodge(width = 0.75)) +
  ggtitle("LGR")


```

#### export in the format needed for stan
```{r}
dplyr::rename(BON_january_spill, BON = january_spill_days) %>% 
  left_join(dplyr::rename(MCN_january_spill, MCN = january_spill_days), by = "year") %>% 
  left_join(dplyr::rename(PRA_january_spill, PRA = january_spill_days), by = "year") %>% 
  left_join(dplyr::rename(RIS_january_spill, RIS = january_spill_days), by = "year") %>% 
  left_join(dplyr::rename(RRE_january_spill, RRE = january_spill_days), by = "year") %>% 
  left_join(dplyr::rename(WEL_january_spill, WEL = january_spill_days), by = "year") %>% 
  left_join(dplyr::rename(ICH_january_spill, ICH = january_spill_days), by = "year") %>% 
  left_join(dplyr::rename(LGR_january_spill, LGR = january_spill_days), by = "year") %>% 
  dplyr::select(-year) %>% 
  mutate(index = row_number()) %>% 
  relocate(index) %>% 
  column_to_rownames("index") -> january_spill_df

# add an empty column for the first state (that doesn't get a temperature effect)
# make sure that the order is the same is the order of states
january_spill_df %>% 
  mutate(MOUTH = 0) %>% 
  relocate(MOUTH) -> january_spill_df
  
dplyr::rename(BON_february_spill, BON = february_spill_days) %>% 
  left_join(dplyr::rename(MCN_february_spill, MCN = february_spill_days), by = "year") %>% 
  left_join(dplyr::rename(PRA_february_spill, PRA = february_spill_days), by = "year") %>% 
  left_join(dplyr::rename(RIS_february_spill, RIS = february_spill_days), by = "year") %>% 
  left_join(dplyr::rename(RRE_february_spill, RRE = february_spill_days), by = "year") %>% 
  left_join(dplyr::rename(WEL_february_spill, WEL = february_spill_days), by = "year") %>% 
  left_join(dplyr::rename(ICH_february_spill, ICH = february_spill_days), by = "year") %>% 
  left_join(dplyr::rename(LGR_february_spill, LGR = february_spill_days), by = "year") %>% 
  dplyr::select(-year) %>% 
  mutate(index = row_number()) %>% 
  relocate(index) %>% 
  column_to_rownames("index") -> february_spill_df

# add an empty column for the first state (that doesn't get a temperature effect)
# make sure that the order is the same is the order of states
february_spill_df %>% 
  mutate(MOUTH = 0) %>% 
  relocate(MOUTH) -> february_spill_df

dplyr::rename(BON_march_spill, BON = march_spill_days) %>% 
  left_join(dplyr::rename(MCN_march_spill, MCN = march_spill_days), by = "year") %>% 
  left_join(dplyr::rename(PRA_march_spill, PRA = march_spill_days), by = "year") %>% 
  left_join(dplyr::rename(RIS_march_spill, RIS = march_spill_days), by = "year") %>% 
  left_join(dplyr::rename(RRE_march_spill, RRE = march_spill_days), by = "year") %>% 
  left_join(dplyr::rename(WEL_march_spill, WEL = march_spill_days), by = "year") %>% 
  left_join(dplyr::rename(ICH_march_spill, ICH = march_spill_days), by = "year") %>% 
  left_join(dplyr::rename(LGR_march_spill, LGR = march_spill_days), by = "year") %>% 
  dplyr::select(-year) %>% 
  mutate(index = row_number()) %>% 
  relocate(index) %>% 
  column_to_rownames("index") -> march_spill_df

# add an empty column for the first state (that doesn't get a temperature effect)
# make sure that the order is the same is the order of states
march_spill_df %>% 
  mutate(MOUTH = 0) %>% 
  relocate(MOUTH) -> march_spill_df

dplyr::rename(BON_april_spill, BON = april_spill_days) %>% 
  left_join(dplyr::rename(MCN_april_spill, MCN = april_spill_days), by = "year") %>% 
  left_join(dplyr::rename(PRA_april_spill, PRA = april_spill_days), by = "year") %>% 
  left_join(dplyr::rename(RIS_april_spill, RIS = april_spill_days), by = "year") %>% 
  left_join(dplyr::rename(RRE_april_spill, RRE = april_spill_days), by = "year") %>% 
  left_join(dplyr::rename(WEL_april_spill, WEL = april_spill_days), by = "year") %>% 
  left_join(dplyr::rename(ICH_april_spill, ICH = april_spill_days), by = "year") %>% 
  left_join(dplyr::rename(LGR_april_spill, LGR = april_spill_days), by = "year") %>% 
  dplyr::select(-year) %>% 
  mutate(index = row_number()) %>% 
  relocate(index) %>% 
  column_to_rownames("index") -> april_spill_df

# add an empty column for the first state (that doesn't get a temperature effect)
# make sure that the order is the same is the order of states
april_spill_df %>% 
  mutate(MOUTH = 0) %>% 
  relocate(MOUTH) -> april_spill_df

write.csv(january_spill_df, here::here("covariate_data", "for_model", "january_spill_df_for_stan.csv"))
write.csv(february_spill_df, here::here("covariate_data", "for_model", "february_spill_df_for_stan.csv"))
write.csv(march_spill_df, here::here("covariate_data", "for_model", "march_spill_df_for_stan.csv"))
write.csv(april_spill_df, here::here("covariate_data", "for_model", "april_spill_df_for_stan.csv"))


```
