==> R/26_year_data_visualization.Rmd <==
---
title: "26_year_data_exploration"
author: "Markus Min"
date: "2023-07-10"
output: html_document
---


#### Description
This Rmd will explore year-to-year variability in movement rates to explore whether year effects should be expected/included in our model.

**Note that this updated Rmd now only contains wild, Upper Columbia steelhead for consistency with the modeling results.**


```{r libraries-setup, echo = FALSE, message = FALSE, warning = FALSE}
knitr::opts_chunk$set(echo = FALSE, message = FALSE, warning = FALSE, fig.height = 10, fig.width = 8)

# load libraries
library(here)
library(janitor)
library(tidyverse)
library(lubridate)
library(car)
library(ggthemes)
library(viridis)
library(ggpubr)
```


```{r load_data}
# so, I think that the most up to date files are in the same folders as the model runs
snake_adults_states_complete <- read.csv(here::here("stan_actual", "deteff_ESU_models", "500iter_3chain_runs", "snake", "snake_adults_states_complete.csv"), row.names = 1)
lowcol_adults_states_complete <- read.csv(here::here("stan_actual", "deteff_ESU_models", "lower_columbia", "lower_columbia_adults_states_complete.csv"), row.names = 1)
midcol_adults_states_complete <- read.csv(here::here("stan_actual", "deteff_ESU_models", "500iter_3chain_runs", "middle_columbia", "middle_columbia_adults_states_complete.csv"), row.names = 1)
uppcol_adults_states_complete <- read.csv(here::here("stan_actual", "deteff_ESU_models", "500iter_3chain_runs", "upper_columbia", "upper_columbia_adults_states_complete.csv"), row.names = 1)

# combine them all
snake_adults_states_complete %>% 
  bind_rows(., midcol_adults_states_complete) %>% 
  bind_rows(., uppcol_adults_states_complete) %>% 
  bind_rows(., lowcol_adults_states_complete) %>% 
  mutate(date_time = ymd_hms(date_time)) -> ASC

# now add tag code metadata, for natal origins
origin_numeric <- data.frame(natal_origin = c("Asotin_Creek", 
                                        "Clearwater_River",
                                        "Deschutes_River", 
                                        "Entiat_River", 
                                        "Fifteenmile_Creek", 
                                        "Grande_Ronde_River", 
                                        "Hood_River",
                                        "Imnaha_River",
                                        "John_Day_River", 
                                        "Methow_River", 
                                        "Okanogan_River", 
                                        "Salmon_River", 
                                        "Tucannon_River", 
                                        "Umatilla_River",
                                        "Walla_Walla_River",
                                        "Wenatchee_River", 
                                        "Yakima_River"),
                             natal_origin_numeric = seq(1,17,1))

origin_rear_actual <- read.csv(here::here("stan_actual", "origin_rear_actual.csv"), row.names = 1)
origin_rear_actual %>% 
  dplyr::rename(natal_origin_numeric = natal_origin) %>% 
  left_join(origin_numeric, by = "natal_origin_numeric") %>% 
  dplyr::select(tag_code_2, natal_origin) -> tag_code_origins

ASC %>% 
  left_join(., tag_code_origins, by = "tag_code_2") -> ASC

# also note which ESU they're from, for plotting
ESU_origins <- data.frame(natal_origin = unique(ASC$natal_origin),
                          ESU = c(rep("Snake", 6),
                                  rep("Middle Columbia", 6),
                                  rep("Upper Columbia", 4),
                                  "Lower Columbia"))

ASC %>% 
  left_join(., ESU_origins, by = "natal_origin") -> ASC


# introduce the loss state
ASC %>% 
  mutate(prev_tag_code = lag(tag_code_2)) %>% 
  mutate(next_tag_code = lead(tag_code_2)) %>% 
  mutate(next_state = lead(state)) %>% 
  mutate(next_state = ifelse(next_tag_code != tag_code_2, "loss", next_state)) -> ASC 

# add run year - takes a long time

# get run year df for joining
run_year <- c("04/05", "05/06", "06/07", "07/08", "08/09", "09/10", "10/11", "11/12", "12/13", "13/14", "14/15", "15/16", "16/17", "17/18", "18/19", "19/20", "20/21","21/22", "22/23")
run_year_start <- seq(ymd_hms("2004-06-01 00:00:00"), ymd_hms("2022-06-01 00:00:00"), by = "years")
run_year_end <- seq(ymd_hms("2005-05-31 23:59:59"), ymd_hms("2023-05-31 23:59:59"), by = "years")
run_year_numeric = seq(4, 22, 1)
run_year_df <- data.frame(run_year, run_year_start, run_year_end, run_year_numeric)

# join run year
ASC %>% 
    rowwise() %>%
    dplyr::mutate(run_year = subset(run_year_df, run_year_start <= date_time & run_year_end >= date_time)$run_year) -> ASC

# add info on run type
tag_code_metadata <- read.csv(here::here("covariate_data", "tag_code_metadata.csv"))

ASC %>% 
  left_join(., dplyr::select(tag_code_metadata, tag_code, rear_type_code), by = "tag_code") %>% 
  mutate(rear_type = ifelse(rear_type_code %in% c("U", "H"), "H", "W"))-> ASC 


```

```{r temp_data}
# get the temp data and add for comparison
read.csv(here::here("covariate_data", "for_model", "temp", "temp_mod_est.csv")) %>% 
  mutate(date = ymd(date)) %>% 
  rowwise() %>%
  dplyr::mutate(run_year = subset(run_year_df, run_year_start <= date & run_year_end >= date)$run_year) %>% 
  dplyr::mutate(run_year_numeric = subset(run_year_df, run_year_start <= date & run_year_end >= date)$run_year_numeric) %>% 
  # keep only run years up to 21/22 (last year of full data)
  filter(run_year_numeric <= 21) -> temp_data

# summarise each into run year averages
temp_data %>% 
  group_by(run_year_numeric) %>% 
  summarise(avg_temp = mean(BON)) -> BON_annual_temp

temp_data %>% 
  group_by(run_year_numeric) %>% 
  summarise(avg_temp = mean(MCN)) -> MCN_annual_temp

temp_data %>% 
  group_by(run_year_numeric) %>% 
  summarise(avg_temp = mean(PRA)) -> PRA_annual_temp

temp_data %>% 
  group_by(run_year_numeric) %>% 
  summarise(avg_temp = mean(RIS)) -> RIS_annual_temp

temp_data %>% 
  group_by(run_year_numeric) %>% 
  summarise(avg_temp = mean(RRE)) -> RRE_annual_temp

temp_data %>% 
  group_by(run_year_numeric) %>% 
  summarise(avg_temp = mean(ICH)) -> ICH_annual_temp

temp_data %>% 
  group_by(run_year_numeric) %>% 
  summarise(avg_temp = mean(WEL)) -> WEL_annual_temp

temp_data %>% 
  group_by(run_year_numeric) %>% 
  summarise(avg_temp = mean(LGR)) -> LGR_annual_temp


```



```{r, fig.height = 5, fig.width = 7}
ASC %>% 
  mutate(run_year_numeric = as.numeric(sub("\\/.*", "", run_year))) %>% 
  group_by(run_year_numeric, ESU) %>% 
  summarise(count = n()) %>% 
  filter(ESU != "Lower Columbia") %>% 
  dplyr::rename(DPS = ESU) -> DPS_counts

# ggplot(DPS_counts, aes(x = run_year_numeric, y = count, fill = DPS))+
#   geom_bar(stat = "identity") +
#   scale_fill_tableau(palette = "Tableau 10") +
#   xlab("Run Year")
```



```{r functions}
# function to reformat data to show movements by year out of a state
state_movement_summary <- function(from_state, DPS = "all DPSs", origin = "all origins", rear_type = "all rear types"){
  if (DPS == "all DPSs" & rear_type == "all rear types"){
    ASC %>% 
      filter(state == from_state) -> one_state_df
  } else if (DPS == "all DPSs" & rear_type != "all rear types"){
    ASC %>% 
      filter(state == from_state) %>% 
      filter(rear_type == rear_type)-> one_state_df
  } else if (DPS != "all DPSs" & rear_type == "all rear types"){
    ASC %>% 
      filter(state == from_state) %>% 
      filter(ESU == DPS)-> one_state_df
    
  } else {
    ASC %>% 
      filter(state == from_state) %>% 
      filter(rear_type == rear_type & ESU == DPS) -> one_state_df
    
  }
  
  # filter by origin (optional)
  if(origin != "all origins"){
    one_state_df %>% 
      filter(natal_origin == origin) -> one_state_df
  }
  

  
  # summarize this by year
  one_state_df %>% 
      # combine the upstream and mouth states for easier visualization
    mutate(next_state = sub(" Mouth| Upstream", "", next_state)) %>% 
    mutate(run_year_numeric = as.numeric(sub("\\/.*", "", run_year))) %>% 
    group_by(run_year_numeric, next_state) %>% 
    summarise(count = n()) %>% 
    mutate(total = sum(count)) %>% 
    mutate(prop = count/total) %>% 
    ungroup() -> movement_counts
  
  
  
  
  return(movement_counts)
}

# also need to be able to visualize sample sizes

# function to combine temp, sample size, and prop plots
movements_counts_aux_df <- function(movement_counts, temp_data){
 movement_counts %>% 
  filter(!(duplicated(run_year_numeric))) %>% 
  left_join(., temp_data, by = "run_year_numeric") -> aux_df
  
  return(aux_df)
  
}


# function to plot movements out of a state

state_movement_summary_plot <- function(movement_counts, movement_counts_aux_df, state_name){
  prop_plot <- ggplot(movement_counts, aes(x = run_year_numeric, y = prop, color = next_state)) +
    geom_point() +
    geom_line() +
    scale_color_tableau(palette = "Tableau 10") +
    # guides(color=guide_legend(nrow=3,byrow=TRUE)) +
    xlab("Run Year")
  
  N_plot <- ggplot(movement_counts_aux_df, aes(x = run_year_numeric, y = total)) +
    geom_bar(stat = "identity") +
    ylab("N")

  T_plot <-ggplot(movement_counts_aux_df, aes(x = run_year_numeric, y = avg_temp)) +
    geom_line() +
    ylab("Mean T (C)")

  comb_plot <- ggarrange(plotlist = list(T_plot + rremove("xlab") + rremove("x.text"), N_plot +  rremove("xlab") + rremove("x.text"), prop_plot),
          nrow = 3, ncol = 1, common.legend = TRUE, legend = "bottom", heights = c(2,2,6))
   comb_plot <- annotate_figure(comb_plot, top = text_grob(state_name, 
               size = 14))
  
  return(comb_plot)
  
}

# a number of confounding things when interpreting this - presence/absence of detection sites, e.g. at Deschutes,
# not accounting for detection efficiency, and different sample sizes of different DPS/origins
# fish in different years, hatchery vs. wild

# check out JDR, Wild
# BON_to_MCN_movement_summary_JDR_W <- state_movement_summary(from_state = "mainstem, BON to MCN", origin = "John_Day_River", rear_type = "W")
# BON_to_MCN_movement_summary_JDR_W_aux <- movements_counts_aux_df(movement_counts = BON_to_MCN_movement_summary_JDR_W, temp_data = MCN_annual_temp)
# 
# BON_to_MCN_movement_summary_JDR_W_plot <- state_movement_summary_plot(movement_counts = BON_to_MCN_movement_summary_JDR_W,
#                                                                       movement_counts_aux_df = BON_to_MCN_movement_summary_JDR_W_aux,
#                                                                       state_name = "Mainstem, BON to MCN")

# ggsave(here::here("figures", "year_plots", "JDR_W_comb_plot.pdf"), plot = JDR_W_comb_plot, height = 10, width = 8)




```

# Plots for all mainstem states
```{r}
# mainstem, BON_to_MCN
BON_to_MCN_movement_summary <- state_movement_summary(DPS = "Upper Columbia", rear_type = "W", from_state = "mainstem, BON to MCN")
BON_to_MCN_movement_summary_aux <- movements_counts_aux_df(movement_counts = BON_to_MCN_movement_summary, temp_data = BON_annual_temp)

BON_to_MCN_movement_summary_plot <- state_movement_summary_plot(movement_counts = BON_to_MCN_movement_summary,
                                                                      movement_counts_aux_df = BON_to_MCN_movement_summary_aux,
                                                                state_name = "Mainstem, BON to MCN")

BON_to_MCN_movement_summary_plot
```

*Note that the Deschutes River Mouth array was only active from 2013-2019.*

```{r}

# mainstem, MCN_to_ICH_or_PRA
MCN_to_ICH_or_PRA_movement_summary <- state_movement_summary(DPS = "Upper Columbia", rear_type = "W", from_state = "mainstem, MCN to ICH or PRA")
MCN_to_ICH_or_PRA_movement_summary_aux <- movements_counts_aux_df(movement_counts = MCN_to_ICH_or_PRA_movement_summary, temp_data = MCN_annual_temp)

MCN_to_ICH_or_PRA_movement_summary_plot <- state_movement_summary_plot(movement_counts = MCN_to_ICH_or_PRA_movement_summary,
                                                                      movement_counts_aux_df = MCN_to_ICH_or_PRA_movement_summary_aux,
                                                                      state_name = "Mainstem, MCN to ICH or PRA")

MCN_to_ICH_or_PRA_movement_summary_plot
```


```{r}
# mainstem, PRA_to_RIS
PRA_to_RIS_movement_summary <- state_movement_summary(DPS = "Upper Columbia", rear_type = "W", from_state = "mainstem, PRA to RIS")
PRA_to_RIS_movement_summary_aux <- movements_counts_aux_df(movement_counts = PRA_to_RIS_movement_summary, temp_data = PRA_annual_temp)

PRA_to_RIS_movement_summary_plot <- state_movement_summary_plot(movement_counts = PRA_to_RIS_movement_summary,
                                                                      movement_counts_aux_df = PRA_to_RIS_movement_summary_aux,
                                                                state_name = "Mainstem, PRA to RIS")

PRA_to_RIS_movement_summary_plot
```

```{r}
# mainstem, RIS_to_RRE
RIS_to_RRE_movement_summary <- state_movement_summary(DPS = "Upper Columbia", rear_type = "W", from_state = "mainstem, RIS to RRE")
RIS_to_RRE_movement_summary_aux <- movements_counts_aux_df(movement_counts = RIS_to_RRE_movement_summary, temp_data = RIS_annual_temp)

RIS_to_RRE_movement_summary_plot <- state_movement_summary_plot(movement_counts = RIS_to_RRE_movement_summary,
                                                                      movement_counts_aux_df = RIS_to_RRE_movement_summary_aux,
                                                                state_name = "Mainstem, RIS to RRE")

RIS_to_RRE_movement_summary_plot
```

#### Wenatchee River only
```{r}
# mainstem, RIS_to_RRE - Wenatchee only
RIS_to_RRE_movement_summary_wenatchee <- state_movement_summary(DPS = "Upper Columbia", origin = "Wenatchee_River", rear_type = "W", from_state = "mainstem, RIS to RRE")
RIS_to_RRE_movement_summary_wenatchee_aux <- movements_counts_aux_df(movement_counts = RIS_to_RRE_movement_summary_wenatchee, temp_data = RIS_annual_temp)

RIS_to_RRE_movement_summary_wenatchee_plot <- state_movement_summary_plot(movement_counts = RIS_to_RRE_movement_summary_wenatchee,
                                                                      movement_counts_aux_df = RIS_to_RRE_movement_summary_wenatchee_aux,
                                                                state_name = "Mainstem, RIS to RRE")

RIS_to_RRE_movement_summary_wenatchee_plot
```


*Note that many arrays in the Wenatchee River came online between 2008 and 2010.*

```{r}
# mainstem, RRE_to_WEL
RRE_to_WEL_movement_summary <- state_movement_summary(DPS = "Upper Columbia", rear_type = "W", from_state = "mainstem, RRE to WEL")
RRE_to_WEL_movement_summary_aux <- movements_counts_aux_df(movement_counts = RRE_to_WEL_movement_summary, temp_data = MCN_annual_temp)

RRE_to_WEL_movement_summary_plot <- state_movement_summary_plot(movement_counts = RRE_to_WEL_movement_summary,
                                                                      movement_counts_aux_df = RRE_to_WEL_movement_summary_aux,
                                                                state_name = "Mainstem, RRE to WEL")

RRE_to_WEL_movement_summary_plot
```

```{r}
# mainstem, upstream_WEL
upstream_WEL_movement_summary <- state_movement_summary(DPS = "Upper Columbia", rear_type = "W", from_state = "mainstem, upstream of WEL")
upstream_WEL_movement_summary_aux <- movements_counts_aux_df(movement_counts = upstream_WEL_movement_summary, temp_data = WEL_annual_temp)

upstream_WEL_movement_summary_plot <- state_movement_summary_plot(movement_counts = upstream_WEL_movement_summary,
                                                                      movement_counts_aux_df = upstream_WEL_movement_summary_aux,
                                                                  state_name = "Mainstem, upstream of WEL")

upstream_WEL_movement_summary_plot
```

*Note that this pattern is driven primarily by the availability of arrays in the tributaries. The Methow River array near the mouth comes online in 2009, and the Okanogan River array near the mouth comes online in 2013. There are some other upstream arrays available for both tributaries in earlier years.*

```{r}
# mainstem, ICH_to_LGR
ICH_to_LGR_movement_summary <- state_movement_summary(DPS = "Upper Columbia", rear_type = "W", from_state = "mainstem, ICH to LGR")
ICH_to_LGR_movement_summary_aux <- movements_counts_aux_df(movement_counts = ICH_to_LGR_movement_summary, temp_data = ICH_annual_temp)

ICH_to_LGR_movement_summary_plot <- state_movement_summary_plot(movement_counts = ICH_to_LGR_movement_summary,
                                                                      movement_counts_aux_df = ICH_to_LGR_movement_summary_aux,
                                                                state_name = "Mainstem, ICH to LGR")

ICH_to_LGR_movement_summary_plot
```

```{r}
# mainstem, upstream_LGR
upstream_LGR_movement_summary <- state_movement_summary(DPS = "Upper Columbia", rear_type = "W", from_state = "mainstem, upstream of LGR")
upstream_LGR_movement_summary_aux <- movements_counts_aux_df(movement_counts = upstream_LGR_movement_summary, temp_data = LGR_annual_temp)

upstream_LGR_movement_summary_plot <- state_movement_summary_plot(movement_counts = upstream_LGR_movement_summary,
                                                                      movement_counts_aux_df = upstream_LGR_movement_summary_aux,
                                                                  state_name = "Mainstem, upstream of LGR")

upstream_LGR_movement_summary_plot
```

*Note that most Snake River tribuatries have poor detection efficiency.*
