==> R/27_spill_effects_visualization.Rmd <==
---
title: "27_spill_effects_visualization"
author: "Markus Min"
date: "2023-07-13"
output: html_document
---


### Description
This Rmd will investigate patterns of fallback with spill, to examine what the data that's going into the model looks like.

**Note that this Rmd only contains wild, Upper Columbia steelhead for consistency with the modeling results**

```{r libraries-setup, echo = FALSE, message = FALSE, warning = FALSE}
knitr::opts_chunk$set(echo = FALSE, message = FALSE, warning = FALSE, fig.height = 10, fig.width = 8)

# load libraries
library(here)
library(janitor)
library(tidyverse)
library(lubridate)
library(car)
library(ggthemes)
library(viridis)
library(ggpubr)
```

```{r load_data}
# so, I think that the most up to date files are in the same folders as the model runs
snake_adults_states_complete <- read.csv(here::here("stan_actual", "deteff_ESU_models", "500iter_3chain_runs", "snake", "snake_adults_states_complete.csv"), row.names = 1)
lowcol_adults_states_complete <- read.csv(here::here("stan_actual", "deteff_ESU_models", "lower_columbia", "lower_columbia_adults_states_complete.csv"), row.names = 1)
midcol_adults_states_complete <- read.csv(here::here("stan_actual", "deteff_ESU_models", "500iter_3chain_runs", "middle_columbia", "middle_columbia_adults_states_complete.csv"), row.names = 1)
uppcol_adults_states_complete <- read.csv(here::here("stan_actual", "deteff_ESU_models", "500iter_3chain_runs", "upper_columbia", "upper_columbia_adults_states_complete.csv"), row.names = 1)

# combine them all
snake_adults_states_complete %>% 
  bind_rows(., midcol_adults_states_complete) %>% 
  bind_rows(., uppcol_adults_states_complete) %>% 
  bind_rows(., lowcol_adults_states_complete) %>% 
  mutate(date_time = ymd_hms(date_time)) -> ASC

# now add tag code metadata, for natal origins
origin_numeric <- data.frame(natal_origin = c("Asotin_Creek", 
                                        "Clearwater_River",
                                        "Deschutes_River", 
                                        "Entiat_River", 
                                        "Fifteenmile_Creek", 
                                        "Grande_Ronde_River", 
                                        "Hood_River",
                                        "Imnaha_River",
                                        "John_Day_River", 
                                        "Methow_River", 
                                        "Okanogan_River", 
                                        "Salmon_River", 
                                        "Tucannon_River", 
                                        "Umatilla_River",
                                        "Walla_Walla_River",
                                        "Wenatchee_River", 
                                        "Yakima_River"),
                             natal_origin_numeric = seq(1,17,1))

origin_rear_actual <- read.csv(here::here("stan_actual", "origin_rear_actual.csv"), row.names = 1)
origin_rear_actual %>% 
  dplyr::rename(natal_origin_numeric = natal_origin) %>% 
  left_join(origin_numeric, by = "natal_origin_numeric") %>% 
  dplyr::select(tag_code_2, natal_origin) -> tag_code_origins

ASC %>% 
  left_join(., tag_code_origins, by = "tag_code_2") -> ASC

# also note which ESU they're from, for plotting
ESU_origins <- data.frame(natal_origin = unique(ASC$natal_origin),
                          ESU = c(rep("Snake", 6),
                                  rep("Middle Columbia", 6),
                                  rep("Upper Columbia", 4),
                                  "Lower Columbia"))

ASC %>% 
  left_join(., ESU_origins, by = "natal_origin") -> ASC


# introduce the loss state
ASC %>% 
  mutate(prev_tag_code = lag(tag_code_2)) %>% 
  mutate(next_tag_code = lead(tag_code_2)) %>% 
  # add in next state, including loss, based on whether the next tag code is the same or not
  mutate(next_state = ifelse(next_tag_code != tag_code_2, "loss", lead(state))) %>% 
  # add in the next date (and if it's loss, NA)
  mutate(date_time_2 = ifelse(next_state == "loss", NA, ymd_hms(format(as_datetime(lead(date_time)))))) -> ASC
  


# Create new variables - window_date_time_1 and window_date_time_2
# For every movement that's not implicit, this will be the same as date_time and date_time_2
# For every implicit movement, this will be the previous and next time that the fish was actually seen, to create the widest possible window
ASC %>% 
  mutate(implicit_movement_prev_date_time = ymd_hms(ifelse(pathway == "implicit" & lag(pathway != "implicit"), format(as_datetime(lag(date_time))), NA_character_))) %>%
  fill(implicit_movement_prev_date_time, .direction = "down") %>%
  mutate(implicit_movement_next_date_time = ymd_hms(ifelse(pathway == "implicit" & lead(pathway != "implicit"), format(as_datetime(lead(date_time))), NA_character_))) %>%
  fill(implicit_movement_next_date_time, .direction = "up") %>%
  # drop all of the rows that you don't need an implicit movement date tim
  mutate(implicit_movement_prev_date_time = ifelse(pathway != "implicit", NA, ymd_hms(format(as_datetime(implicit_movement_prev_date_time))))) %>%
  mutate(implicit_movement_next_date_time = ifelse(pathway != "implicit", NA, ymd_hms(format(as_datetime(implicit_movement_next_date_time))))) %>% 
  # dplyr::select(tag_code, state, next_state, pathway, date_time, date_time_2, implicit_movement_prev_date_time, implicit_movement_next_date_time)
  mutate(window_date_time_1 = ymd_hms(format(as_datetime(ifelse(pathway == "implicit", implicit_movement_prev_date_time, date_time))))) %>% 
  mutate(window_date_time_2 = ymd_hms(format(as_datetime(ifelse(pathway == "implicit", implicit_movement_next_date_time, date_time_2))))) -> ASC
  # dplyr::select(tag_code, state, next_state, pathway, date_time, date_time_2, implicit_movement_prev_date_time, implicit_movement_next_date_time, window_date_time_1, window_date_time_2)

  # The below block is actually a bad idea, because since fallback is almost always implicit, 
  # this change makes it such that the window for fallback is basically nothing
  # MAJOR CHANGE - how current date_time is calculated
  # because for our purposes now we want to know the entire duration of the possible window
  # of time that a fish was in a state, we will have change the treatment of date time
  # for implicit site visits. In the previous scripts, it's such that the date time
  # for any implicit site visits is the next time we see the fish (which is good for some purposes)
  # but for this, we want it now to be the previous time we saw the fish - this gives 
  # the full possible window
  # mutate(implicit_movement_date_time = ymd_hms(ifelse(pathway == "implicit" & lag(pathway != "implicit"), format(as_datetime(lag(date_time))), NA_character_))) %>% 
  # fill(implicit_movement_date_time) %>% 
  # mutate(implicit_movement_date_time = ifelse(pathway != "implicit", NA, implicit_movement_date_time)) %>% 
  # mutate(date_time = format(as_datetime(ifelse(pathway == "implicit", implicit_movement_date_time, date_time)))) %>% 
  # # give it an NA for time if it's loss
  # mutate(date_time_2 = ifelse(next_state == "loss", NA, format(as_datetime(lead(date_time))))) %>% 
  # mutate(date_time = ymd_hms(date_time)) %>% 
  # mutate(date_time_2 = ymd_hms(date_time_2)) %>% 
  # dplyr::select(-implicit_movement_date_time) -> ASC 

# add run year - takes a long time

# get run year df for joining
run_year <- c("04/05", "05/06", "06/07", "07/08", "08/09", "09/10", "10/11", "11/12", "12/13", "13/14", "14/15", "15/16", "16/17", "17/18", "18/19", "19/20", "20/21","21/22", "22/23")
run_year_start <- seq(ymd_hms("2004-06-01 00:00:00"), ymd_hms("2022-06-01 00:00:00"), by = "years")
run_year_end <- seq(ymd_hms("2005-05-31 23:59:59"), ymd_hms("2023-05-31 23:59:59"), by = "years")
run_year_numeric = seq(4, 22, 1)
run_year_df <- data.frame(run_year, run_year_start, run_year_end, run_year_numeric)

# join run year
ASC %>% 
    rowwise() %>%
    dplyr::mutate(run_year = subset(run_year_df, run_year_start <= date_time & run_year_end >= date_time)$run_year) -> ASC

# add info on run type
tag_code_metadata <- read.csv(here::here("covariate_data", "tag_code_metadata.csv"))

ASC %>% 
  left_join(., dplyr::select(tag_code_metadata, tag_code, rear_type_code), by = "tag_code") %>% 
  mutate(rear_type = ifelse(rear_type_code %in% c("U", "H"), "H", "W"))-> ASC 

# Keep only upper columbia, wild
ASC %>% 
  filter(ESU == "Upper Columbia" & rear_type == "W") %>% 
  # also drop okanogan, because they get dropped in the model due to small sample size
  filter(natal_origin != "Okanogan_River") %>% 
  # also extract just the date, so it can be joined
  mutate(date = date(date_time)) -> UCW_ASC


```

```{r load_spill_data}
spill_window_data <- read.csv(here::here("covariate_data", "for_model", "window_spill_for_stan.csv"), row.names = 1)
jan_spill_days <- read.csv(here::here("covariate_data", "for_model", "january_spill_df_for_stan.csv"), row.names = 1)
feb_spill_days <- read.csv(here::here("covariate_data", "for_model", "february_spill_df_for_stan.csv"), row.names = 1)
mar_spill_days <- read.csv(here::here("covariate_data", "for_model", "march_spill_df_for_stan.csv"), row.names = 1)
apr_spill_days <- read.csv(here::here("covariate_data", "for_model", "april_spill_df_for_stan.csv"), row.names = 1)

# need to add dates back into window spill for combining with movement data
spill_window_data$date <- seq(ymd("2005-06-01"), ymd("2022-12-31"), by = "days")
spill_window_data %>% 
  relocate(date) -> spill_window_data

# need to add run years back into days of spill for combining with movement data
feb_spill_days$run_year <- c("04/05", "05/06", "06/07", "07/08", "08/09", "09/10", "10/11", "11/12", "12/13", "13/14", "14/15", "15/16", "16/17", "17/18", "18/19", "19/20", "20/21","21/22")
feb_spill_days %>% 
  relocate(run_year) -> feb_spill_days

mar_spill_days$run_year <- c("04/05", "05/06", "06/07", "07/08", "08/09", "09/10", "10/11", "11/12", "12/13", "13/14", "14/15", "15/16", "16/17", "17/18", "18/19", "19/20", "20/21","21/22")
mar_spill_days %>% 
  relocate(run_year) -> mar_spill_days

apr_spill_days$run_year <- c("04/05", "05/06", "06/07", "07/08", "08/09", "09/10", "10/11", "11/12", "12/13", "13/14", "14/15", "15/16", "16/17", "17/18", "18/19", "19/20", "20/21","21/22")
apr_spill_days %>% 
  relocate(run_year) -> apr_spill_days

jan_spill_days$run_year <- c("04/05", "05/06", "06/07", "07/08", "08/09", "09/10", "10/11", "11/12", "12/13", "13/14", "14/15", "15/16", "16/17", "17/18", "18/19", "19/20", "20/21","21/22")
jan_spill_days %>% 
  relocate(run_year) -> jan_spill_days


# add the spill window data to the ASC
UCW_ASC %>% 
  left_join(., spill_window_data, by = "date") -> UCW_ASC_spill
```



## En-route fallback

#### Fallback over Bonneville

```{r}
UCW_ASC_spill %>% 
  filter(state == "mainstem, BON to MCN") %>% 
  mutate(fallback = ifelse(next_state == "mainstem, mouth to BON", 1, 0))-> UCW_ASC_spill_BON_to_MCN

table(UCW_ASC_spill_BON_to_MCN$next_state)
# table(UCW_ASC_spill_BON_to_MCN$fallback)

BON_spill_fallback_glm <- glm(fallback ~ BON, data = UCW_ASC_spill_BON_to_MCN, family = binomial)


BON_fallback_plot <- ggplot(UCW_ASC_spill_BON_to_MCN, aes(x = BON, y = fallback)) +
  geom_point() +
  xlab("Spill Volume at Bonneville Dam (over median travel time window)") +
  stat_smooth(method = "glm", method.args = list(family = binomial))

# also plot a histogram of distribution of spill
BON_spill_hist <- ggplot(UCW_ASC_spill_BON_to_MCN, aes(x = BON)) +
  geom_histogram() +
  xlab("Spill Volume at Bonneville Dam (over median travel time window)")

ggarrange(plotlist = list(BON_fallback_plot + rremove("xlab") + rremove("x.text"), BON_spill_hist),
          nrow = 2, heights = c(6,4))
```


#### Fallback over Priest Rapids

```{r}
UCW_ASC_spill %>% 
  filter(state == "mainstem, PRA to RIS") %>% 
  mutate(fallback = ifelse(next_state == "mainstem, MCN to ICH or PRA", 1, 0))-> UCW_ASC_spill_PRA_to_RIS

table(UCW_ASC_spill_PRA_to_RIS$next_state)
# table(UCW_ASC_spill_PRA_to_RIS$fallback)

PRA_spill_fallback_glm <- glm(fallback ~ PRA, data = UCW_ASC_spill_PRA_to_RIS, family = binomial)


PRA_fallback_plot <- ggplot(UCW_ASC_spill_PRA_to_RIS, aes(x = PRA, y = fallback)) +
  geom_point() +
  xlab("Spill Volume at Priest Rapids Dam (over median travel time window)") +
  stat_smooth(method = "glm", method.args = list(family = binomial))

# also plot a histogram of distribution of spill
PRA_spill_hist <- ggplot(UCW_ASC_spill_PRA_to_RIS, aes(x = PRA)) +
  geom_histogram() +
  xlab("Spill Volume at Priest Rapids Dam (over median travel time window)")

ggarrange(plotlist = list(PRA_fallback_plot + rremove("xlab") + rremove("x.text"), PRA_spill_hist),
          nrow = 2, heights = c(6,4))
```


## Post-overshoot fallback

#### How correlated are number of spill days in different months?

```{r load_spill_days_data}
knitr::opts_chunk$set(echo = FALSE, message = FALSE, warning = FALSE, fig.height = 6, fig.width = 8)

BON_april_spill <- read.csv(here::here("covariate_data", "for_model", "spill", "BON_april_spill.csv"))
MCN_april_spill <- read.csv(here::here("covariate_data", "for_model", "spill", "MCN_april_spill.csv"))
PRA_april_spill <- read.csv(here::here("covariate_data", "for_model", "spill", "PRA_april_spill.csv"))
RIS_april_spill <- read.csv(here::here("covariate_data", "for_model", "spill", "RIS_april_spill.csv"))
RRE_april_spill <- read.csv(here::here("covariate_data", "for_model", "spill", "RRE_april_spill.csv"))
WEL_april_spill <- read.csv(here::here("covariate_data", "for_model", "spill", "WEL_april_spill.csv"))
ICH_april_spill <- read.csv(here::here("covariate_data", "for_model", "spill", "ICH_april_spill.csv"))
LGR_april_spill <- read.csv(here::here("covariate_data", "for_model", "spill", "LGR_april_spill.csv"))

BON_march_spill <- read.csv(here::here("covariate_data", "for_model", "spill", "BON_march_spill.csv"))
MCN_march_spill <- read.csv(here::here("covariate_data", "for_model", "spill", "MCN_march_spill.csv"))
PRA_march_spill <- read.csv(here::here("covariate_data", "for_model", "spill", "PRA_march_spill.csv"))
RIS_march_spill <- read.csv(here::here("covariate_data", "for_model", "spill", "RIS_march_spill.csv"))
RRE_march_spill <- read.csv(here::here("covariate_data", "for_model", "spill", "RRE_march_spill.csv"))
WEL_march_spill <- read.csv(here::here("covariate_data", "for_model", "spill", "WEL_march_spill.csv"))
ICH_march_spill <- read.csv(here::here("covariate_data", "for_model", "spill", "ICH_march_spill.csv"))
LGR_march_spill <- read.csv(here::here("covariate_data", "for_model", "spill", "LGR_march_spill.csv"))

BON_february_spill <- read.csv(here::here("covariate_data", "for_model", "spill", "BON_february_spill.csv"))
MCN_february_spill <- read.csv(here::here("covariate_data", "for_model", "spill", "MCN_february_spill.csv"))
PRA_february_spill <- read.csv(here::here("covariate_data", "for_model", "spill", "PRA_february_spill.csv"))
RIS_february_spill <- read.csv(here::here("covariate_data", "for_model", "spill", "RIS_february_spill.csv"))
RRE_february_spill <- read.csv(here::here("covariate_data", "for_model", "spill", "RRE_february_spill.csv"))
WEL_february_spill <- read.csv(here::here("covariate_data", "for_model", "spill", "WEL_february_spill.csv"))
ICH_february_spill <- read.csv(here::here("covariate_data", "for_model", "spill", "ICH_february_spill.csv"))
LGR_february_spill <- read.csv(here::here("covariate_data", "for_model", "spill", "LGR_february_spill.csv"))

BON_january_spill <- read.csv(here::here("covariate_data", "for_model", "spill", "BON_january_spill.csv"))
MCN_january_spill <- read.csv(here::here("covariate_data", "for_model", "spill", "MCN_january_spill.csv"))
PRA_january_spill <- read.csv(here::here("covariate_data", "for_model", "spill", "PRA_january_spill.csv"))
RIS_january_spill <- read.csv(here::here("covariate_data", "for_model", "spill", "RIS_january_spill.csv"))
RRE_january_spill <- read.csv(here::here("covariate_data", "for_model", "spill", "RRE_january_spill.csv"))
WEL_january_spill <- read.csv(here::here("covariate_data", "for_model", "spill", "WEL_january_spill.csv"))
ICH_january_spill <- read.csv(here::here("covariate_data", "for_model", "spill", "ICH_january_spill.csv"))
LGR_january_spill <- read.csv(here::here("covariate_data", "for_model", "spill", "LGR_january_spill.csv"))
```


```{r plot_spill_vif}
spill_days_vif <- function(monthly_spill){
  # reformat for VIF 
  monthly_spill %>% 
    pivot_wider(names_from = name, values_from = value) -> monthly_spill_wide
  
  # run multiple linear regressions and calculate VIF
  jan_spill_mlm <- lm(january_spill_days ~ february_spill_days + march_spill_days + april_spill_days, data = monthly_spill_wide)
  jan_spill_vif <- vif(jan_spill_mlm)

  feb_spill_mlm <- lm(february_spill_days ~ january_spill_days + march_spill_days + april_spill_days, data = monthly_spill_wide)
  feb_spill_vif <- vif(feb_spill_mlm)
  
  mar_spill_mlm <- lm(march_spill_days ~ january_spill_days + february_spill_days + april_spill_days, data = monthly_spill_wide)
  mar_spill_vif <- vif(mar_spill_mlm)
  
  apr_spill_mlm <- lm(april_spill_days ~ january_spill_days + february_spill_days + march_spill_days, data = monthly_spill_wide)
  apr_spill_vif <- vif(apr_spill_mlm)
  
  return(list(jan_spill_vif, feb_spill_vif, mar_spill_vif, apr_spill_vif))
} 

```

``` {r}
# make a quick plot for spill availability
BON_march_spill %>% 
  left_join(., BON_april_spill, by = "year") %>% 
  left_join(., BON_february_spill, by = "year") %>% 
  left_join(., BON_january_spill, by = "year") %>% 
  pivot_longer(., cols = c("april_spill_days", "march_spill_days", "february_spill_days", "january_spill_days")) -> BON_monthly_spill

ggplot(BON_monthly_spill, aes(x = year, y = value, color = name)) +
  geom_point(position = position_dodge(width = 0.75)) +
  ggtitle("BON")

# Bonneville doesn't matter because it's not an overshoot fallback movement for any origins
# spill_days_vif(monthly_spill = BON_monthly_spill)
```

``` {r}
MCN_march_spill %>% 
  left_join(., MCN_april_spill, by = "year") %>% 
  left_join(., MCN_february_spill, by = "year") %>% 
  left_join(., MCN_january_spill, by = "year") %>% 
  pivot_longer(., cols = c("april_spill_days", "march_spill_days", "february_spill_days", "january_spill_days")) -> MCN_monthly_spill

ggplot(MCN_monthly_spill, aes(x = year, y = value, color = name)) +
  geom_point(position = position_dodge(width = 0.75)) +
  ggtitle("MCN")

spill_days_vif(monthly_spill = MCN_monthly_spill)
```

``` {r}
PRA_march_spill %>% 
  left_join(., PRA_april_spill, by = "year") %>% 
  left_join(., PRA_february_spill, by = "year") %>% 
  left_join(., PRA_january_spill, by = "year") %>% 
  pivot_longer(., cols = c("april_spill_days", "march_spill_days", "february_spill_days", "january_spill_days")) -> PRA_monthly_spill

ggplot(PRA_monthly_spill, aes(x = year, y = value, color = name)) +
  geom_point(position = position_dodge(width = 0.75)) +
  ggtitle("PRA")

spill_days_vif(monthly_spill = PRA_monthly_spill)
```

``` {r}
RIS_march_spill %>% 
  left_join(., RIS_april_spill, by = "year") %>% 
  left_join(., RIS_february_spill, by = "year") %>% 
  left_join(., RIS_january_spill, by = "year") %>% 
  pivot_longer(., cols = c("april_spill_days", "march_spill_days", "february_spill_days", "january_spill_days")) -> RIS_monthly_spill

ggplot(RIS_monthly_spill, aes(x = year, y = value, color = name)) +
  geom_point(position = position_dodge(width = 0.75)) +
  ggtitle("RIS")

spill_days_vif(monthly_spill = RIS_monthly_spill)
```

``` {r}
RRE_march_spill %>% 
  left_join(., RRE_april_spill, by = "year") %>% 
  left_join(., RRE_february_spill, by = "year") %>% 
  left_join(., RRE_january_spill, by = "year") %>% 
  pivot_longer(., cols = c("april_spill_days", "march_spill_days", "february_spill_days", "january_spill_days")) -> RRE_monthly_spill

ggplot(RRE_monthly_spill, aes(x = year, y = value, color = name)) +
  geom_point(position = position_dodge(width = 0.75)) +
  ggtitle("RRE")

spill_days_vif(monthly_spill = RRE_monthly_spill)
```

``` {r}
WEL_march_spill %>% 
  left_join(., WEL_april_spill, by = "year") %>% 
  left_join(., WEL_february_spill, by = "year") %>% 
  left_join(., WEL_january_spill, by = "year") %>% 
  pivot_longer(., cols = c("april_spill_days", "march_spill_days", "february_spill_days", "january_spill_days")) -> WEL_monthly_spill

WEL_march_spill %>% 
  left_join(., WEL_april_spill, by = "year") %>% 
  left_join(., WEL_february_spill, by = "year") %>% 
  left_join(., WEL_january_spill, by = "year") -> WEL_monthly_spill_wide

ggplot(WEL_monthly_spill, aes(x = year, y = value, color = name)) +
  geom_point(position = position_dodge(width = 0.75)) +
  ggtitle("WEL")

spill_days_vif(monthly_spill = WEL_monthly_spill)

# why are these parameter estimates so weird? Maybe it has to do with correlations?
# VIF is quite low but still

cor(dplyr::select(WEL_monthly_spill_wide, -year))

cor(WEL_monthly_spill_wide$january_spill_days + WEL_monthly_spill_wide$march_spill_days, WEL_monthly_spill_wide$february_spill_days + WEL_monthly_spill_wide$april_spill_days)

# right now jan and march have a positive effect, feb and april have a negative effect
ggplot(WEL_monthly_spill_wide, aes(x = january_spill_days + march_spill_days, y = february_spill_days + april_spill_days)) +
  geom_point() +
  ggtitle("Wells Dam, Spill Days")
```

``` {r}
ICH_march_spill %>% 
  left_join(., ICH_april_spill, by = "year") %>% 
  left_join(., ICH_february_spill, by = "year") %>% 
  left_join(., ICH_january_spill, by = "year") %>% 
  pivot_longer(., cols = c("april_spill_days", "march_spill_days", "february_spill_days", "january_spill_days")) -> ICH_monthly_spill

ggplot(ICH_monthly_spill, aes(x = year, y = value, color = name)) +
  geom_point(position = position_dodge(width = 0.75)) +
  ggtitle("ICH")

spill_days_vif(monthly_spill = ICH_monthly_spill)
```

``` {r}
LGR_march_spill %>% 
  left_join(., LGR_april_spill, by = "year") %>% 
  left_join(., LGR_february_spill, by = "year") %>% 
  left_join(., LGR_january_spill, by = "year") %>% 
  pivot_longer(., cols = c("april_spill_days", "march_spill_days", "february_spill_days", "january_spill_days")) -> LGR_monthly_spill

ggplot(LGR_monthly_spill, aes(x = year, y = value, color = name)) +
  geom_point(position = position_dodge(width = 0.75)) +
  ggtitle("LGR")

spill_days_vif(monthly_spill = LGR_monthly_spill)
```



### Explore parameter estimates from model versus data on fallback

- Eight post-overshoot fallback parameters are estimated for Upper Columbia Wild Steelhead: Fallback over Wells Dam (for Entiat and Wenatchee Steelhead) and fallback over Rocky Reach Dam (for Entiat Steelhead), in four months (January, February, March, and April)

The pattern for fallback at Rocky Reach is more in line with our hypotheses: Spill in February, March, and April is weakly correlated with post-overshoot fallback; January Spill is weakly negative correlated (but this credible interval is much wider).

The pattern for fallback at Wells Dam is confusing: it oscillates between being positive and negative, with much smaller credible intervals: January and March have a weakly positive correlation with fallback, February and April a weakly negative correlation.

```{r}
### functions to visualize fallback versus spill days for a given month and a given dam

# Function 1: filter and refomat data
fallback_v_spilldays_reformat <- function(movement_data, spill_data, spill_month_numeric, origins, overshoot_state, fallback_state, dam_abbrev){
  movement_data %>% 
    left_join(., spill_data, by = "run_year") %>% 
    filter(natal_origin %in% origins) %>% 
    filter(state == overshoot_state) %>% 
    mutate(fallback = ifelse(next_state == fallback_state, 1, 0)) %>% 
    mutate(month_1 = month(window_date_time_1)) %>% 
    mutate(month_2 = month(window_date_time_2)) %>% 
    mutate(spill_month_overlap = ifelse(month_1 > month_2,
                        ifelse(spill_month_numeric %in% c(seq(month_1, 12), seq(1, month_2)), 1, 0),
                  ifelse(spill_month_numeric %in% seq(month_1, month_2), 1, 0)))  %>% 
    # if the next state is loss, month_2 will be NA - so give it spill_month_overlap
    mutate(spill_month_overlap = ifelse(is.na(month_2), 1, spill_month_overlap)) %>% 
    filter(spill_month_overlap == 1) %>% 
    dplyr::select(tag_code, state, next_state, natal_origin, window_date_time_1, window_date_time_2, run_year, fallback, month_1, month_2, spill_month_overlap, dam_abbrev) %>% 
    # this rename call allows us to use the next function
    dplyr::rename(spill_days = dam_abbrev) -> fallback_v_spilldays_data
  
  # return the reformatted data
  return(fallback_v_spilldays_data)

}

# Function 2: Generate the plot
fallback_v_spilldays_plot <- function(fallback_v_spilldays_data, month_name, dam_name) {
  # fit a logistic regression to spill days as predictor and predict for vals
  monthspill_fallback_glm <- glm(fallback ~ spill_days, data = fallback_v_spilldays_data, family = binomial)
  dam_monthspill_predict <- data.frame(spill_days = seq(0,31,1))
  dam_monthspill_predict$fallback <- predict(monthspill_fallback_glm, newdata = dam_monthspill_predict, type = "response")

  # summarise for the plot
  fallback_v_spilldays_data %>% 
    group_by(spill_days) %>% 
    summarise(not_fallback = n() - sum(fallback), fallback = sum(fallback))  %>% 
    pivot_longer(cols = c(not_fallback, fallback)) %>% 
    dplyr::rename(fallback = name) %>% 
    mutate(fallback = ifelse(fallback == "fallback", 1, 0)) -> fallback_v_spilldays_count
    
  # generate the plot
  fallback_v_spilldays_plot <- ggplot(fallback_v_spilldays_count, aes(x = spill_days, y = fallback)) +
    geom_point(aes(size = value)) +
    xlab(paste0(month_name, " spill days at ", dam_name)) +
    geom_line(data = dam_monthspill_predict, lty = 2)
  
  # return the plot and the glm
  return(list(fallback_v_spilldays_plot, summary(monthspill_fallback_glm)))

}

```

### Spill days for combinations of spill month and dam

#### January Spill, Rocky Reach Dam
```{r}
jan_RRE <- fallback_v_spilldays_reformat(movement_data = UCW_ASC, 
                                         spill_data = jan_spill_days, 
                                         spill_month_numeric = 1, 
                                         origins = c("Wenatchee_River"), 
                                         overshoot_state = "mainstem, RRE to WEL", 
                                         fallback_state = "mainstem, RIS to RRE", 
                                         dam_abbrev = "RRE")

# Function 2: Generate the plot
fallback_v_spilldays_plot(jan_RRE, month_name = "January", dam_name = "Rocky Reach Dam")

setdiff(jan_RRE$tag_code, fish_tag_codes_script)
```
#### February Spill, Rocky Reach Dam
```{r}
feb_RRE <- fallback_v_spilldays_reformat(movement_data = UCW_ASC, 
                                         spill_data = feb_spill_days, 
                                         spill_month_numeric = 2, 
                                         origins = c("Wenatchee_River"), 
                                         overshoot_state = "mainstem, RRE to WEL", 
                                         fallback_state = "mainstem, RIS to RRE", 
                                         dam_abbrev = "RRE")

# Function 2: Generate the plot
fallback_v_spilldays_plot(feb_RRE, month_name = "February", dam_name = "Rocky Reach Dam")
```

#### March Spill, Rocky Reach Dam
```{r}
mar_RRE <- fallback_v_spilldays_reformat(movement_data = UCW_ASC, 
                                         spill_data = mar_spill_days, 
                                         spill_month_numeric = 3, 
                                         origins = c("Wenatchee_River"), 
                                         overshoot_state = "mainstem, RRE to WEL", 
                                         fallback_state = "mainstem, RIS to RRE", 
                                         dam_abbrev = "RRE")

# Function 2: Generate the plot
fallback_v_spilldays_plot(mar_RRE, month_name = "March", dam_name = "Rocky Reach Dam")
```

#### April Spill, Rocky Reach Dam
```{r}
apr_RRE <- fallback_v_spilldays_reformat(movement_data = UCW_ASC, 
                                         spill_data = apr_spill_days, 
                                         spill_month_numeric = 4, 
                                         origins = c("Wenatchee_River"), 
                                         overshoot_state = "mainstem, RRE to WEL", 
                                         fallback_state = "mainstem, RIS to RRE", 
                                         dam_abbrev = "RRE")

# Function 2: Generate the plot
fallback_v_spilldays_plot(apr_RRE, month_name = "April", dam_name = "Rocky Reach Dam")
```


#### January Spill, Wells Dam
```{r}
jan_WEL <- fallback_v_spilldays_reformat(movement_data = UCW_ASC, 
                                         spill_data = jan_spill_days, 
                                         spill_month_numeric = 1, 
                                         origins = c("Wenatchee_River", "Entiat_River"), 
                                         overshoot_state = "mainstem, upstream of WEL", 
                                         fallback_state = "mainstem, RRE to WEL", 
                                         dam_abbrev = "WEL")

# Function 2: Generate the plot
fallback_v_spilldays_plot(jan_WEL, month_name = "January", dam_name = "Wells Dam")
```
#### February Spill, Wells Dam
```{r}
feb_WEL <- fallback_v_spilldays_reformat(movement_data = UCW_ASC, 
                                         spill_data = feb_spill_days, 
                                         spill_month_numeric = 2, 
                                         origins = c("Wenatchee_River", "Entiat_River"), 
                                         overshoot_state = "mainstem, upstream of WEL", 
                                         fallback_state = "mainstem, RRE to WEL", 
                                         dam_abbrev = "WEL")

# Function 2: Generate the plot
fallback_v_spilldays_plot(feb_WEL, month_name = "February", dam_name = "Wells Dam")
```

#### March Spill, Wells Dam
```{r}
mar_WEL <- fallback_v_spilldays_reformat(movement_data = UCW_ASC, 
                                         spill_data = mar_spill_days, 
                                         spill_month_numeric = 3, 
                                         origins = c("Wenatchee_River", "Entiat_River"), 
                                         overshoot_state = "mainstem, upstream of WEL", 
                                         fallback_state = "mainstem, RRE to WEL", 
                                         dam_abbrev = "WEL")

# Function 2: Generate the plot
fallback_v_spilldays_plot(mar_WEL, month_name = "March", dam_name = "Wells Dam")
```

#### April Spill, Wells Dam
```{r}
apr_WEL <- fallback_v_spilldays_reformat(movement_data = UCW_ASC, 
                                         spill_data = apr_spill_days, 
                                         spill_month_numeric = 4, 
                                         origins = c("Wenatchee_River", "Entiat_River"), 
                                         overshoot_state = "mainstem, upstream of WEL", 
                                         fallback_state = "mainstem, RRE to WEL", 
                                         dam_abbrev = "WEL")

# Function 2: Generate the plot
fallback_v_spilldays_plot(apr_WEL, month_name = "April", dam_name = "Wells Dam")
```

```{r}
# explore for Wells Dam, how many fish are getting each predictor
dplyr::rename(dplyr::select(jan_WEL, tag_code, spill_month_overlap), January = spill_month_overlap) %>% 
full_join(dplyr::rename(dplyr::select(feb_WEL, tag_code, spill_month_overlap), February = spill_month_overlap), by = "tag_code") %>% 
full_join(dplyr::rename(dplyr::select(mar_WEL, tag_code, spill_month_overlap), March = spill_month_overlap), by = "tag_code") %>% 
  full_join(dplyr::rename(dplyr::select(apr_WEL, tag_code, spill_month_overlap), April = spill_month_overlap)) %>% 
  replace(is.na(.), 0) -> WEL_spill_month_predictors
  
WEL_spill_month_predictors %>% 
  mutate(months = paste0(ifelse(January == 1, "Jan", ""), 
                         ifelse(February == 1, ", Feb", ""), 
                         ifelse(March == 1, ", Mar", ""), 
                         ifelse(April == 1, ", Apr", ""))) %>% 
  mutate(months = ifelse(months == ", Apr", "Apr", months)) -> WEL_spill_month_predictors 
  
table(WEL_spill_month_predictors$months)
```
